Convert mediation_data to a single data.frame

It can be helpful to combine all the data in a mediation_data S4 object into a single data.frame. This concatenates all data sources horizontally, into an N samples x (all outcomes, mediators, treatments, ...) matrix.

bind_mediation(exper)

Arguments

exper

An object of class mediation_data containing the variables that we want horizontally concatenated.

Value

A data.frame containing all variables from the original mediation_data object.

Examples

exper <- demo_joy() |>
  mediation_data("PHQ", "treatment", starts_with("ASV"))
exper
#> [Mediation Data] 
#> 100 samples with measurements for, 
#> 1 treatment: treatment 
#> 5 mediators: ASV1, ASV2, ... 
#> 1 outcome: PHQ 
bind_mediation(exper)
#> # A tibble: 100 × 7
#>       ASV1    ASV2   ASV3   ASV4     ASV5    PHQ treatment
#>      <dbl>   <dbl>  <dbl>  <dbl>    <dbl>  <dbl> <fct>    
#>  1  0.862   0.429  -0.830 -0.579  1.78    -0.784 Treatment
#>  2 -0.243   0.122  -0.504 -0.145  0.387   -0.953 Treatment
#>  3 -0.206  -1.14   -1.19   0.526 -0.919    1.79  Control  
#>  4  0.0192 -0.558  -0.752  1.73  -1.58     0.349 Treatment
#>  5  0.0296  1.05    1.46   1.45  -0.0841   0.259 Treatment
#>  6  0.550   0.678  -0.829  1.52  -2.09    -0.806 Control  
#>  7 -2.27    0.0385  0.290 -0.384  0.00357  0.106 Treatment
#>  8  2.68   -0.356  -0.480  1.83  -0.356   -0.334 Treatment
#>  9 -0.361   0.783  -0.605 -0.551  1.15     1.64  Treatment
#> 10  0.213   0.804   1.46  -0.866 -0.221   -0.644 Control  
#> # ℹ 90 more rows
  
exper <- demo_spline(tau = c(2, 1)) |>
  mediation_data(starts_with("outcome"), "treatment", "mediator")
exper
#> [Mediation Data] 
#> 5000 samples with measurements for, 
#> 1 treatment: treatment 
#> 1 mediator: mediator 
#> 2 outcomes: outcome_1, outcome_2 
bind_mediation(exper)
#> # A tibble: 5,000 × 4
#>    mediator outcome_1 outcome_2 treatment
#>       <dbl>     <dbl>     <dbl> <fct>    
#>  1  -3.00      -0.830   -0.209  0        
#>  2  -1.82      -0.799   -0.0772 0        
#>  3  -0.796     -1.54    -1.63   0        
#>  4  -2.08      -0.658    0.190  0        
#>  5  -1.29      -0.352    0.790  0        
#>  6  -0.0596    -0.795   -0.634  0        
#>  7  -2.72      -0.588    0.290  0        
#>  8  -3.71      -0.401    0.525  0        
#>  9  -0.430     -0.982   -0.716  0        
#> 10  -0.0279    -0.749   -0.575  0        
#> # ℹ 4,990 more rows