Bayesian Regression Model across Responses

Apply a Bayesian regression model in parallel across each response $y$ in an outcome or mediation model. This can be helpful when we want to share information across related

brms_model(...)

Arguments

...

Keyword parameters passed to brm..

Value

model An object of class model with estimator, predictor, and sampler functions associated wtih a Bayesian regression model.

See also

Examples

exper <- demo_joy() |>
  mediation_data("PHQ", "treatment", starts_with("ASV"))
multimedia(exper, brms_model()) |>
  estimate(exper)
#> Compiling Stan program...
#> Trying to compile a simple C file
#> Running /Library/Frameworks/R.framework/Resources/bin/R CMD SHLIB foo.c
#> sh: gcc-14.0.3: command not found
#> -using SDK: ‘MacOSX13.3.sdk’
#> \gcc-14.0.3 -I"/Library/Frameworks/R.framework/Resources/include" -DNDEBUG   -I/opt/R/arm64/include    -fPIC  -mtune=native -g -O2 -Wall -pedantic -Wconversion -c foo.c -o foo.o
#> |/bin/sh: gcc-14.0.3: command not found
#> /make: *** [foo.o] Error 127
#> -\ 
#> Start sampling
#> [Multimedia Analysis] 
#> Treatments: treatment 
#> Outcomes: PHQ 
#> Mediators: ASV1, ASV2, ... 
#>   
#> [Models] 
#> mediation: A fitted lm_model(). 
#> outcome: A fitted brms_model(). 

# example with another dataset
exper <- demo_spline(tau = c(2, 1)) |>
  mediation_data(starts_with("outcome"), "treatment", "mediator")
fit <- multimedia(exper, brms_model()) |>
  estimate(exper)
#> Compiling Stan program...
#> Trying to compile a simple C file
#> Running /Library/Frameworks/R.framework/Resources/bin/R CMD SHLIB foo.c
#> sh: gcc-14.0.3: command not found
#> -using SDK: ‘MacOSX13.3.sdk’
#> \gcc-14.0.3 -I"/Library/Frameworks/R.framework/Resources/include" -DNDEBUG   -I"/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/Rcpp/include/"  -I"/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/RcppEigen/include/"  -I"/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/RcppEigen/include/unsupported"  -I"/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/BH/include" -I"/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/StanHeaders/include/src/"  -I"/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/StanHeaders/include/"  -I"/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/RcppParallel/include/"  -I"/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/rstan/include" -DEIGEN_NO_DEBUG  -DBOOST_DISABLE_ASSERTS  -DBOOST_PENDING_INTEGER_LOG2_HPP  -DSTAN_THREADS  -DUSE_STANC3 -DSTRICT_R_HEADERS  -DBOOST_PHOENIX_NO_VARIADIC_EXPRESSION  -D_HAS_AUTO_PTR_ETC=0  -include '/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/StanHeaders/include/stan/math/prim/fun/Eigen.hpp'  -D_REENTRANT -DRCPP_PARALLEL_USE_TBB=1   -I/opt/R/arm64/include    -fPIC  -mtune=native -g -O2 -Wall -pedantic -Wconversion -c foo.c -o foo.o
#> |/bin/sh: gcc-14.0.3: command not found
#> /make: *** [foo.o] Error 127
#> -\ 
#> Start sampling
#> Start sampling
fit@outcome
#> A fitted brms_model(). 
#> $outcome_1
#>  Family: gaussian 
#>   Links: mu = identity; sigma = identity 
#> Formula: outcome_1 ~ mediator + treatment 
#>    Data: data (Number of observations: 5000) 
#>   Draws: 1 chains, each with iter = 2000; warmup = 1000; thin = 1;
#>          total post-warmup draws = 1000
#> 
#> Population-Level Effects: 
#>            Estimate Est.Error l-95% CI u-95% CI Rhat Bulk_ESS Tail_ESS
#> Intercept     -0.12      0.02    -0.15    -0.08 1.00      927      704
#> mediator       0.20      0.01     0.18     0.21 1.00      760      675
#> treatment1     2.23      0.03     2.17     2.28 1.00      865      782
#> 
#> Family Specific Parameters: 
#>       Estimate Est.Error l-95% CI u-95% CI Rhat Bulk_ESS Tail_ESS
#> sigma     0.83      0.01     0.81     0.85 1.00      922      699
#> 
#> Draws were sampled using sampling(NUTS). For each parameter, Bulk_ESS
#> and Tail_ESS are effective sample size measures, and Rhat is the potential
#> scale reduction factor on split chains (at convergence, Rhat = 1).
#> 
#> $outcome_2
#>  Family: gaussian 
#>   Links: mu = identity; sigma = identity 
#> Formula: outcome_2 ~ mediator + treatment 
#>    Data: data (Number of observations: 5000) 
#>   Draws: 1 chains, each with iter = 2000; warmup = 1000; thin = 1;
#>          total post-warmup draws = 1000
#> 
#> Population-Level Effects: 
#>            Estimate Est.Error l-95% CI u-95% CI Rhat Bulk_ESS Tail_ESS
#> Intercept     -0.03      0.02    -0.06     0.01 1.00      922      716
#> mediator       0.24      0.01     0.23     0.25 1.00      920      793
#> treatment1     1.04      0.03     0.99     1.10 1.00      908      767
#> 
#> Family Specific Parameters: 
#>       Estimate Est.Error l-95% CI u-95% CI Rhat Bulk_ESS Tail_ESS
#> sigma     0.80      0.01     0.78     0.81 1.01      898      566
#> 
#> Draws were sampled using sampling(NUTS). For each parameter, Bulk_ESS
#> and Tail_ESS are effective sample size measures, and Rhat is the potential
#> scale reduction factor on split chains (at convergence, Rhat = 1).
#>