A Demo Dataset (Spline)

This is a simple dataset with nonlinear relationships between the outcome and mediators. It is used simply to illustrate the design of multimedia. The mediator->outcome effect is generated from a random spline function.

demo_spline(n_samples = 5000, tau = c(2, 2))

Arguments

n_samples

The number of samples to generate in the toy example

tau

The true direct efefcts associated with the two outcomes. Defaults to 2, 2.

Value

xy A tibble whose columns include the treatment, mediation, and outcome variables.

Examples

demo_spline()
#> # A tibble: 5,000 × 4
#>    outcome_1 outcome_2 mediator treatment
#>        <dbl>     <dbl>    <dbl> <fct>    
#>  1    -0.830   -0.209   -3.00   0        
#>  2    -0.799   -0.0772  -1.82   0        
#>  3    -1.54    -1.63    -0.796  0        
#>  4    -0.658    0.190   -2.08   0        
#>  5    -0.352    0.790   -1.29   0        
#>  6    -0.795   -0.634   -0.0596 0        
#>  7    -0.588    0.290   -2.72   0        
#>  8    -0.401    0.525   -3.71   0        
#>  9    -0.982   -0.716   -0.430  0        
#> 10    -0.749   -0.575   -0.0279 0        
#> # ℹ 4,990 more rows
demo_spline(20)
#> # A tibble: 20 × 4
#>    outcome_1 outcome_2 mediator treatment
#>        <dbl>     <dbl>    <dbl> <fct>    
#>  1   -0.973     0.636    -2.87  0        
#>  2   -1.53      0.0710   -2.81  0        
#>  3    1.19      0.656    -1.03  0        
#>  4    1.15     -0.0518    0.721 0        
#>  5   -0.0371   -0.107    -1.32  0        
#>  6    0.0415    2.07     -4.26  0        
#>  7   -0.778     1.19     -3.91  0        
#>  8    0.980    -0.197    -0.493 0        
#>  9   -0.336     1.08     -2.59  0        
#> 10    0.957    -0.222    -0.490 0        
#> 11    1.73      2.93     -2.32  1        
#> 12    3.51      2.42      0.898 1        
#> 13    2.73      1.97      1.29  1        
#> 14    0.0279    2.18      4.62  1        
#> 15    3.13      1.77      0.474 1        
#> 16    1.47      0.360     2.75  1        
#> 17    1.74      0.416     3.04  1        
#> 18    2.36      1.39     -0.667 1        
#> 19    2.01      1.48      1.79  1        
#> 20    0.637    -0.0516    3.89  1        
demo_spline(20, c(0, 10))
#> # A tibble: 20 × 4
#>    outcome_1 outcome_2 mediator treatment
#>        <dbl>     <dbl>    <dbl> <fct>    
#>  1   0.884       1.27  -1.24    0        
#>  2  -0.621      -0.730  2.66    0        
#>  3   0.289       0.176 -0.122   0        
#>  4   0.838       1.28  -1.43    0        
#>  5  -0.522      -1.21  -0.00924 0        
#>  6  -0.324      -1.08  -0.690   0        
#>  7  -1.44       -1.78   2.46    0        
#>  8   0.569       0.743 -1.34    0        
#>  9  -0.123      -0.647 -0.262   0        
#> 10   0.668       1.44  -3.30    0        
#> 11  -0.986       9.69   1.41    1        
#> 12  -1.11        9.19   2.24    1        
#> 13   1.43       10.3    4.02    1        
#> 14   1.44       11.3    3.36    1        
#> 15  -0.944       9.89   1.66    1        
#> 16  -0.543       8.96   2.85    1        
#> 17   0.00692     9.80   0.0481  1        
#> 18  -0.988       9.73   1.97    1        
#> 19  -0.723      10.1    1.31    1        
#> 20   2.19       11.6    4.07    1