Convert a Summarized Experiment to a data.frame

This is a small helepr function to concatenate the colData and assays within SummarizedExperiment objects into a single data.frame. This is not necessary for the essential multimedia workflow, it is only exported for potential independent interest.

exper_df(exper)

Arguments

exper

An object of class SummarizedExperiment.

Examples

demo_joy() |>
  multimedia:::exper_df()
#>            ASV1         ASV2         ASV3         ASV4         ASV5 treatment
#> 1    1.08447878  0.415214482  1.665756226  0.483465501 -1.420159986 Treatment
#> 2   -0.50570230 -0.026531444 -0.319885378 -1.226306596 -0.924715205   Control
#> 3   -0.30472403  0.728016216 -1.340730686  0.200489854  1.383315107 Treatment
#> 4    0.62083630 -0.879021959  0.377883120  1.182295118  1.247228303   Control
#> 5   -0.42001272 -0.739147338 -0.561757361 -0.249377749  0.641846865   Control
#> 6   -0.19239432 -0.006615728 -0.296438477 -0.076578056  0.183993241 Treatment
#> 7   -0.11139846 -0.585257374 -1.746507319 -0.352198427  0.249385302 Treatment
#> 8    0.88743394  0.861457347  0.004022176  0.623301911  1.397416198   Control
#> 9    1.39617421 -1.318480803  0.492717826 -1.969493420 -1.405403073   Control
#> 10   1.14340434  0.179006280 -0.168648995 -0.250092865  1.067921684 Treatment
#> 11   0.34414045  0.773622993  1.701303560  0.807966355  0.398390381 Treatment
#> 12   0.41435101 -0.207968949 -0.436178439  2.212359850 -1.734403936   Control
#> 13   0.77632409  0.068767609  0.219048698 -0.902533324  0.665923010   Control
#> 14   0.31652400  0.418175599  0.883319137 -1.639474136 -1.263018178 Treatment
#> 15   0.11430633 -0.395683328 -1.404788517 -0.068589280 -0.022838681 Treatment
#> 16  -0.58223749 -0.849308756  1.418065106 -0.680986243 -0.088844842   Control
#> 17  -0.42195558 -1.155767936 -0.145172116  0.886248072 -1.286261880 Treatment
#> 18  -1.42640203 -0.375859600 -0.295177923  0.734123986  1.036595310   Control
#> 19   0.48601567  1.262549611 -0.885288490  0.747388899  0.205051472   Control
#> 20  -0.63873080 -1.671862884  0.105365031 -0.111475025 -0.169121753 Treatment
#> 21  -1.42713651  0.574406282  0.385784734  0.583799923  0.426480977 Treatment
#> 22   0.55184965 -1.942258234  0.581930955 -1.663084858 -0.545754382 Treatment
#> 23   0.58048841  0.672763283  0.274909929  0.700722557  0.138433901   Control
#> 24   0.44471626  1.469262248  0.608164993 -0.952303809  1.461287353 Treatment
#> 25   1.53104796  1.646974906  0.247034550 -1.795457593 -0.925848198   Control
#> 26   0.08633602 -1.313646787  0.203397547  0.171575336 -0.379873163   Control
#> 27  -0.31523223 -1.162992737 -1.201830092 -1.216328498  0.025213990 Treatment
#> 28  -1.45515139 -0.454733727  0.148952601  1.641492598  0.452125113   Control
#> 29   1.41873185 -0.273046112 -1.267702846 -0.863497799 -1.137165311 Treatment
#> 30  -0.27024720  0.535802126  1.505417526  0.556473533  0.802291308   Control
#> 31   1.51271252  0.284041750 -0.220439712  1.794047900 -0.019089522   Control
#> 32  -0.13712315  0.406705667 -0.747297429  0.174075518 -0.777260182   Control
#> 33  -1.75442548  0.867832394 -0.736636044  0.294761930 -0.865641184   Control
#> 34  -1.61996739  0.150330183 -0.557608168 -0.370362071  1.500915904 Treatment
#> 35  -0.63993812 -0.194730973 -0.692222168  0.613422673  1.192753153 Treatment
#> 36   0.69947183  0.521512336  0.830442335  0.737189295  0.873988941   Control
#> 37  -1.12814757  0.984134026  0.716241084  0.511984714  1.149954603 Treatment
#> 38   0.45570050  2.207249366  1.574770751  1.911141677  1.372031863 Treatment
#> 39  -0.63982587 -0.251100896  0.535840996  0.240558750 -0.002540616 Treatment
#> 40  -2.08606054 -0.128910661 -0.086803087 -0.420768250 -0.348500451 Treatment
#> 41  -1.20437828 -0.651586500 -1.518608045  1.702521726  1.034860581 Treatment
#> 42  -0.22927651  0.352169858 -1.807237608  0.605134797  0.470008368 Treatment
#> 43   1.43927780  0.053825531 -0.364947655 -0.449272267  0.642911651 Treatment
#> 44   0.66912057  0.669336893  0.651543880 -0.349861590  2.322513890 Treatment
#> 45  -1.64832937  1.158923394 -0.976744767  0.841463285  1.213966577   Control
#> 46  -1.36160048  0.437047903 -0.900787253 -1.334255172 -1.404168945 Treatment
#> 47   1.37250970 -0.080979985 -1.447823793  0.546320112 -0.872922975 Treatment
#> 48  -0.40428489 -0.531037305 -1.174838639 -0.927852177 -1.452928682 Treatment
#> 49  -0.12208980  1.781305065  0.486694427 -0.122623929  0.270038246   Control
#> 50   0.37173612 -1.312021786  0.325020294  0.037209516 -1.035247817   Control
#> 51  -0.46117205 -1.561055991 -1.852344068 -0.606965309  2.030695544 Treatment
#> 52  -0.26873781 -0.489401489  2.180502559 -0.120529398  0.529288090 Treatment
#> 53   0.21897377  0.054279641 -1.232491012  0.146452210 -1.874841917   Control
#> 54  -1.86546170  1.408765998 -0.122685038  0.113256055  2.054666714   Control
#> 55  -0.41731311  0.475089682  0.656812231  0.491839100 -0.571218761 Treatment
#> 56  -1.29851992 -0.178125158 -0.093632074  0.251918044 -1.656166223 Treatment
#> 57   1.28307102  1.175918848 -0.314287946 -0.009618572  0.151002536 Treatment
#> 58  -1.94809433  0.672141278  0.010745691 -0.742342032 -0.725083709   Control
#> 59  -0.78893553 -0.477800634 -2.321567216  1.734068379 -1.292947128   Control
#> 60  -2.22469137  0.510967697 -0.553599967 -0.810194284 -0.770680329   Control
#> 61   0.34110517  2.758498632 -0.266795588  0.963781585 -1.188843578 Treatment
#> 62  -1.72191515 -1.061832358  0.847605478  0.021251964 -2.563306866 Treatment
#> 63  -1.31787188  2.183886008  1.375047953 -0.189376582  0.316731198 Treatment
#> 64  -0.38293988 -2.616820799 -0.457916373 -0.206199931  0.079480124 Treatment
#> 65   0.66686336  0.433324904  1.297923904 -0.730649722  0.005891371 Treatment
#> 66  -1.10739237  1.955851650 -0.746612891  1.593151475 -0.925844925 Treatment
#> 67   0.54558678  0.290608147 -2.019614601 -0.168106438 -1.381358871   Control
#> 68  -0.75479469 -1.344730313  1.158530197 -1.437330650 -0.498270939 Treatment
#> 69   1.13671040  0.411093743 -0.270128318  0.148437631  0.207095705   Control
#> 70   0.24868716  1.229695024 -0.433411372 -0.836342154  0.805297235 Treatment
#> 71  -1.15519880 -0.285152356  0.276465433  0.676065295  0.909342861   Control
#> 72   0.23647254 -3.474151148  0.708182568  0.552799994  0.761591391 Treatment
#> 73   0.38295660  0.578308268 -0.462198994  0.008013421 -1.149454018 Treatment
#> 74   0.23820839 -0.443015117  0.277478113  0.430809061 -0.024149124 Treatment
#> 75   0.07963249  1.394164049 -2.265909586  0.741215839 -0.586079180   Control
#> 76  -1.17602701 -0.571795036 -0.782645342  0.904020046  2.249854469 Treatment
#> 77  -0.14647574 -0.571820168 -2.081254585 -1.395317320 -1.220232291   Control
#> 78  -0.09475448  1.083911967  1.199057070 -1.588518771 -0.619878028   Control
#> 79  -0.10339401  0.032601142  0.266225689  0.440355348 -0.137286504   Control
#> 80   1.76950017 -0.356159939 -1.605273561 -0.502770902  0.219261615 Treatment
#> 81  -0.59122330  0.455609916  0.755151372 -0.540553175  0.163439736 Treatment
#> 82   1.21238420 -0.715822110 -2.989052919  0.331758450 -0.504277308 Treatment
#> 83  -0.59906426 -0.039236143 -0.700402801  0.290493976 -0.420015550   Control
#> 84   0.10654025  0.748670985  0.263854170  0.517053404 -0.668126083 Treatment
#> 85   0.26721235  0.399383536  0.558089488 -0.888891346  0.004571647   Control
#> 86   0.93562166  0.713742638 -0.798724656 -0.199668106 -0.513342578   Control
#> 87   0.37245268 -0.992264557 -0.080704547  0.252548445 -1.453090795   Control
#> 88   0.23928553 -1.194364754  1.016216455  0.052194871  0.291970154 Treatment
#> 89  -1.44176645  0.859978270  0.474435247  0.719911626  0.321916912   Control
#> 90  -1.84689260  0.286598129  2.204435397 -0.097048928  0.981753527   Control
#> 91  -0.65079923 -0.305228608 -0.093978763 -0.594341084  0.004682658 Treatment
#> 92   0.90878951 -0.579979593 -2.183795052 -2.590074316  0.557876145   Control
#> 93   1.02956200  0.829314635 -0.140265720  0.383454528 -0.259642199   Control
#> 94  -2.86117002  0.114962375  1.204247248  2.210561673  0.624417482 Treatment
#> 95  -2.11939461 -0.152394620 -2.068709886  2.166306028 -1.407663597   Control
#> 96   1.03423322 -0.691528255  1.697578565  0.368600241 -0.269071811 Treatment
#> 97  -1.77231629 -0.383642641  2.119915760 -0.937131153  0.880130924 Treatment
#> 98  -0.21393194  0.941165199 -0.653377933 -0.014361657  0.412678983 Treatment
#> 99   0.91825837 -1.078252877  0.650628368  0.391729999  0.560203703 Treatment
#> 100  0.54622536  1.647389546 -0.648465154  1.358462481  0.446260400   Control
#>              PHQ
#> 1    0.468952613
#> 2   -0.506626904
#> 3   -1.561556256
#> 4    1.292147843
#> 5    1.052126125
#> 6   -0.765347220
#> 7   -1.011479667
#> 8   -1.410796776
#> 9   -0.680601847
#> 10   2.171133950
#> 11  -0.129813523
#> 12   1.392235276
#> 13   0.187643966
#> 14   1.592210430
#> 15   2.395871008
#> 16   0.826836725
#> 17  -1.703119604
#> 18   0.368095760
#> 19   1.626522840
#> 20  -0.646050092
#> 21   0.017061308
#> 22   1.348349946
#> 23   0.172925347
#> 24  -1.810590991
#> 25   0.150624563
#> 26  -1.003772117
#> 27  -0.531589975
#> 28  -0.769127955
#> 29  -0.170921137
#> 30   0.506586544
#> 31   0.438425177
#> 32   0.410309283
#> 33   1.719280262
#> 34  -0.482167507
#> 35  -1.612112972
#> 36  -1.033103611
#> 37  -0.626575149
#> 38   1.988694355
#> 39  -0.617107630
#> 40  -0.093954286
#> 41  -1.866776444
#> 42   0.283834221
#> 43  -0.782300340
#> 44   0.346099273
#> 45   0.514091951
#> 46   1.294934858
#> 47   0.352033430
#> 48  -0.193778564
#> 49   0.607246790
#> 50   3.466623216
#> 51   1.273467606
#> 52  -0.004955086
#> 53   0.006056590
#> 54   0.582897838
#> 55  -0.636339068
#> 56   0.478329441
#> 57  -0.516066636
#> 58   1.173761180
#> 59   0.527858040
#> 60  -0.401446329
#> 61   0.274497455
#> 62   0.995934209
#> 63  -0.313310478
#> 64  -1.542884914
#> 65  -1.347816931
#> 66   0.633476808
#> 67   0.306755934
#> 68   0.427162730
#> 69  -0.479869169
#> 70  -0.162113769
#> 71  -1.023374716
#> 72  -0.353598613
#> 73   0.295707379
#> 74  -0.843026548
#> 75   0.093150733
#> 76   0.984684044
#> 77  -0.364512384
#> 78  -0.520444679
#> 79  -0.533331137
#> 80   0.415137951
#> 81   0.264808810
#> 82  -0.707789745
#> 83  -0.141147068
#> 84   0.091252279
#> 85   0.178235796
#> 86  -0.205118355
#> 87   0.662124829
#> 88   1.776724741
#> 89  -0.690619048
#> 90  -0.081422180
#> 91  -0.296774476
#> 92   0.373486180
#> 93   0.303544840
#> 94  -0.852631912
#> 95   1.013231616
#> 96   0.205271080
#> 97   0.086058337
#> 98   0.298300979
#> 99   0.683394336
#> 100 -0.406321263