Overall Indirect Effect

Direct Effects from Estimated Model

indirect_overall(model, exper = NULL, t1 = 1, t2 = 2)

Arguments

model

An object of class multimedia containing the estimated mediation and outcome models whose mediation and outcome predictions we want to compare.

exper

An object of class multimedia_data containing the mediation and outcome data from which the direct effects are to be estimated.

t1

The reference level of the treatment to be used when computing the indirect effect.

t2

The alternative level of the treatment to be used when computing the indirect effect.

Value

A data.frame summarizing the overall indirect effects associated with different settings of j in the equation above.

Details

Estimate direct effects associated with a multimedia model. These estimates are formed using Equation (10) of our preprint:

$$ \frac{1}{2} \sum_{j = 0}^{1} \sum_{i = 1}^{N} \hat{Y}_{i}(t2, \hat{M}(t2)) - \hat{Y}_{i}(t1, \hat{M}(t1)) $$

Examples

# example with null data
exper <- demo_joy() |>
  mediation_data("PHQ", "treatment", starts_with("ASV"))
fit <- multimedia(exper) |>
  estimate(exper)

indirect_overall(fit)
#>   outcome direct_setting            contrast indirect_effect
#> 1     PHQ        Control Control - Treatment    -0.007384257
#> 2     PHQ      Treatment Control - Treatment    -0.007384257

# example with another dataset
exper <- demo_spline(tau = c(2, 1)) |>
  mediation_data(starts_with("outcome"), "treatment", "mediator")
fit <- multimedia(exper) |>
  estimate(exper)
indirect_overall(fit)
#>     outcome direct_setting contrast indirect_effect
#> 1 outcome_1              0    0 - 1      -0.5951599
#> 2 outcome_2              0    0 - 1      -0.8893485
#> 3 outcome_1              1    0 - 1      -0.5951599
#> 4 outcome_2              1    0 - 1      -0.8893485