Logistic Normal Multinomial Model

Apply a logistic normal multinomial model to jointly model a vector of count responses $y$ in an outcome or mediation model. This is a common choice for data where the parameter of interest is the composition across responses (e.g., microbiome).

Usage

lnm_model(...)

Arguments

...

Keyword parameters passed to lnm.

Value

model An object of class model with estimator, predictor, and sampler functions associated wtih a lienar model.

See also

Examples

m <- lnm_model()
mat <- data.frame(matrix(rpois(250, 10), 25, 10))
colnames(mat) <- paste0("y", seq_len(6))
fit <- estimator(m)(y1 + y2 + y3 + y4 ~ y5 + y6, mat)
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 6.1e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.61 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Iteration: 250 / 250 [100%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 0.1].
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100       -32553.812             1.000            1.000
#> Chain 1:    200       -12687.070             1.283            1.566
#> Chain 1:    300        -6146.668             1.210            1.064
#> Chain 1:    400        -1541.103             1.655            1.566
#> Chain 1:    500        -1117.472             1.400            1.064
#> Chain 1:    600         -686.995             1.271            1.064
#> Chain 1:    700         -550.889             1.124            1.000
#> Chain 1:    800         -457.169             1.010            1.000
#> Chain 1:    900         -425.312             0.906            0.627
#> Chain 1:   1000         -405.858             0.820            0.627
#> Chain 1:   1100         -390.443             0.724            0.379   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   1200         -372.678             0.572            0.247   MAY BE DIVERGING... INSPECT ELBO
#> Chain 1:   1300         -364.184             0.468            0.205
#> Chain 1:   1400         -354.054             0.172            0.075
#> Chain 1:   1500         -344.962             0.137            0.048
#> Chain 1:   1600         -334.434             0.077            0.048
#> Chain 1:   1700         -328.332             0.054            0.039
#> Chain 1:   1800         -315.009             0.038            0.039
#> Chain 1:   1900         -308.324             0.033            0.031
#> Chain 1:   2000         -302.110             0.030            0.029
#> Chain 1:   2100         -291.631             0.030            0.029
#> Chain 1:   2200         -286.867             0.027            0.026
#> Chain 1:   2300         -280.676             0.026            0.026
#> Chain 1:   2400         -271.862             0.027            0.026
#> Chain 1:   2500         -269.198             0.025            0.022
#> Chain 1:   2600         -267.029             0.023            0.022
#> Chain 1:   2700         -264.660             0.022            0.022
#> Chain 1:   2800         -256.765             0.021            0.022
#> Chain 1:   2900         -255.964             0.019            0.021
#> Chain 1:   3000         -253.532             0.018            0.017
#> Chain 1:   3100         -247.518             0.017            0.017
#> Chain 1:   3200         -250.718             0.016            0.013
#> Chain 1:   3300         -248.230             0.015            0.010
#> Chain 1:   3400         -244.498             0.013            0.010
#> Chain 1:   3500         -245.646             0.013            0.010
#> Chain 1:   3600         -242.674             0.013            0.012
#> Chain 1:   3700         -242.105             0.013            0.012
#> Chain 1:   3800         -240.015             0.010            0.010
#> Chain 1:   3900         -239.164             0.010            0.010
#> Chain 1:   4000         -239.485             0.010            0.010   MEAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.
#> Warning: Pareto k diagnostic value is 2.05. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.