Sample from the Logistic Normal Multinomial

This samples from the posterior predictive of a fitted logistic-normal multinomial model.

Usage

lnm_sampler(fit, newdata = NULL, indices = NULL, ...)

Arguments

fit

The fitted LNM model from which to draw posterior predictive samples.

newdata

A data.frame containing new inputs from which to sample responses. If NULL, defaults to the data used to estimate fit.

indices

The coordinates of the response from which to draw samples.

...

Additional parameters passed to sample.

Value

y_star A data.frame of samples y associated wtih the new inputs.

Examples

m <- lnm_model()
mat <- data.frame(matrix(rpois(250, 10), 25, 10))
colnames(mat) <- paste0("y", 1:6)
fit <- estimator(m)(y1 + y2 + y3 + y4 ~ y5 + y6, mat)
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 2.3e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.23 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100         -474.307             1.000            1.000
#> Chain 1:    200         -300.472             0.789            1.000
#> Chain 1:    300         -307.290             0.534            0.579
#> Chain 1:    400         -255.619             0.451            0.579
#> Chain 1:    500         -255.664             0.361            0.202
#> Chain 1:    600         -246.963             0.306            0.202
#> Chain 1:    700         -249.069             0.264            0.035
#> Chain 1:    800         -261.984             0.237            0.049
#> Chain 1:    900         -278.793             0.217            0.049
#> Chain 1:   1000         -245.051             0.209            0.060
#> Chain 1:   1100         -275.806             0.121            0.060
#> Chain 1:   1200         -245.508             0.075            0.060
#> Chain 1:   1300         -247.530             0.074            0.060
#> Chain 1:   1400         -248.193             0.054            0.049
#> Chain 1:   1500         -258.166             0.058            0.049
#> Chain 1:   1600         -242.340             0.061            0.060
#> Chain 1:   1700         -274.909             0.072            0.065
#> Chain 1:   1800         -244.683             0.079            0.112
#> Chain 1:   1900         -254.465             0.077            0.112
#> Chain 1:   2000         -244.381             0.067            0.065
#> Chain 1:   2100         -243.351             0.056            0.041
#> Chain 1:   2200         -246.254             0.045            0.039
#> Chain 1:   2300         -241.347             0.046            0.039
#> Chain 1:   2400         -243.177             0.047            0.039
#> Chain 1:   2500         -275.174             0.055            0.041
#> Chain 1:   2600         -254.154             0.056            0.041
#> Chain 1:   2700         -243.205             0.049            0.041
#> Chain 1:   2800         -239.800             0.038            0.038
#> Chain 1:   2900         -239.227             0.035            0.020
#> Chain 1:   3000         -240.618             0.031            0.014
#> Chain 1:   3100         -246.608             0.033            0.020
#> Chain 1:   3200         -244.258             0.033            0.020
#> Chain 1:   3300         -240.424             0.032            0.016
#> Chain 1:   3400         -247.458             0.034            0.024
#> Chain 1:   3500         -259.001             0.027            0.024
#> Chain 1:   3600         -256.524             0.020            0.016
#> Chain 1:   3700         -241.787             0.022            0.016
#> Chain 1:   3800         -241.830             0.020            0.016
#> Chain 1:   3900         -240.983             0.020            0.016
#> Chain 1:   4000         -251.512             0.024            0.024
#> Chain 1:   4100         -243.936             0.025            0.028
#> Chain 1:   4200         -242.715             0.024            0.028
#> Chain 1:   4300         -241.521             0.023            0.028
#> Chain 1:   4400         -240.384             0.021            0.010   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.
#> Warning: Pareto k diagnostic value is 1.28. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.
lnm_sampler(fit, depth = 10)
#>       y1 y2 y3 y4
#>  [1,]  3  3  1  3
#>  [2,]  2  1  4  3
#>  [3,]  3  3  2  2
#>  [4,]  1  7  2  0
#>  [5,]  2  3  2  3
#>  [6,]  3  3  3  1
#>  [7,]  2  1  3  4
#>  [8,]  3  0  4  3
#>  [9,]  3  2  3  2
#> [10,]  2  3  2  3
#> [11,]  5  2  2  1
#> [12,]  3  2  2  3
#> [13,]  2  1  4  3
#> [14,]  3  1  0  6
#> [15,]  5  3  0  2
#> [16,]  0  5  3  2
#> [17,]  2  7  0  1
#> [18,]  2  1  3  4
#> [19,]  2  4  1  3
#> [20,]  3  1  2  4
#> [21,]  1  2  3  4
#> [22,]  0  4  3  3
#> [23,]  0  3  3  4
#> [24,]  3  2  1  4
#> [25,]  3  2  3  2
lnm_sampler(fit, depth = 100)
#>       y1 y2 y3 y4
#>  [1,] 33 18 26 23
#>  [2,] 24 34 23 19
#>  [3,] 25 34 20 21
#>  [4,] 21 14 31 34
#>  [5,] 27 36 20 17
#>  [6,] 28 24 21 27
#>  [7,] 19 20 29 32
#>  [8,] 29 20 28 23
#>  [9,] 17 32 36 15
#> [10,] 19 35 18 28
#> [11,] 22 29 29 20
#> [12,] 26 27 26 21
#> [13,] 14 40 22 24
#> [14,] 20 44 22 14
#> [15,] 22 23 32 23
#> [16,] 26 31 26 17
#> [17,] 19 25 21 35
#> [18,] 20 25 27 28
#> [19,] 29 27 23 21
#> [20,] 16 27 39 18
#> [21,] 16 23 23 38
#> [22,] 19 24 33 24
#> [23,] 25 21 32 22
#> [24,] 24 24 24 28
#> [25,] 26 25 21 28