Access Class-Specific Mediators

This is a shorthand for accessing mediator-related slots in classes exported by the multimedia package.

Usage

mediators(object)

Arguments

object

An object whose mediators we want to access (either their names or values).

Value

The output will depend on the class of the object that is passed in. For multimedia objects, this returns a character vector of mediators. For mediation data objects, this returns the mediator mediator data.

Examples

exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
mediators(exper)
#>            ASV1        ASV2        ASV3         ASV4        ASV5
#> 1    0.33335462 -1.06946473 -0.71464483 -0.388499711 -0.96292350
#> 2    0.20105967  1.50566422 -0.27877289 -0.128347889  0.80191317
#> 3    1.70216981 -0.88022053 -0.76681413 -0.487367805 -0.62202156
#> 4    0.13698532 -0.83466211  0.33296842  1.581062100  0.66665550
#> 5   -1.35080293  0.43093420 -1.23031055  0.018458931  1.91877875
#> 6    2.62774794  0.10982537 -0.21421582  0.020412507  1.78771067
#> 7   -0.74891673  0.45934684  0.36822394 -2.164413069  0.14261736
#> 8    1.13467583  0.19955800  0.53142430 -0.444158251 -0.02017365
#> 9   -1.21364560 -1.59135421  1.25057916 -1.060556747  0.14008874
#> 10  -1.47461760 -0.09214321  0.85561932  0.918600116 -0.08114791
#> 11   0.48920784 -1.11754556  1.35027727 -1.191968350  1.54896366
#> 12  -0.09451211  0.73660789  1.24572420  0.520207772 -1.57768631
#> 13  -1.77729721 -1.25612412 -0.05935078 -0.416045086  0.42470270
#> 14   0.95627185 -1.56288813  1.18844357  0.645963900 -1.34857380
#> 15  -0.13950089 -0.03241454 -0.24436205  0.411656698  1.20933983
#> 16   0.11705034  0.22985404  1.16247320 -1.088729573  0.40305765
#> 17  -0.41119051  0.85751920 -3.23964725 -0.599798640  0.43029836
#> 18   0.66119027 -0.36840907  0.47845487  0.843113596  0.29034391
#> 19   1.76910879 -0.26336339 -0.62592964  3.052477434 -2.77177963
#> 20  -0.10364266  0.32056717 -0.36322005 -0.331107414  1.52261461
#> 21  -0.69720610 -0.61558626  0.69215311 -0.704623997 -1.09236521
#> 22   0.28708928  0.27230139  1.42300857  1.323047557 -0.73175535
#> 23   1.22816552 -0.36011424  0.24956601  2.207385456  1.60129582
#> 24   0.02391378 -0.07686883 -1.32388770  2.005330195  0.58184950
#> 25  -0.31332833  0.33829098 -0.39098663 -0.491239595 -0.98769909
#> 26  -0.15343940 -1.68263424  0.63463242  0.717518567 -0.59420582
#> 27  -2.61255002 -0.75598609 -0.20966913 -1.392665326 -0.49379444
#> 28   0.80127684  0.13671421 -1.49331772 -0.406259192  0.36877436
#> 29  -0.43453364  0.03349820  1.27056870 -0.558247276 -1.87117085
#> 30  -0.58693039 -0.39891887 -0.37025710  0.354052161  0.28542832
#> 31   0.81298859  0.56954264  0.54344593 -1.094762987  0.08851209
#> 32  -2.24992633  0.02493277 -0.02470053  0.527587060  3.13722597
#> 33  -0.44854590 -0.11799398 -1.61557449  1.555090056  0.78170937
#> 34  -0.25711631  1.33071567  0.42249917  0.755647032  0.10919727
#> 35   1.34365849  0.30417199 -2.22390387  0.613564640 -4.09800626
#> 36   1.90982507  0.62284131 -0.61055604 -0.385188705 -0.24991918
#> 37   1.34817396 -0.70064928  2.12161395  0.745335321  0.67028790
#> 38  -0.91379182 -0.61592554  0.28021404  1.363208321 -0.62166474
#> 39   1.41425991  0.19963882 -2.29662023  0.813478886  1.06347014
#> 40   0.67796908  2.44644242 -0.62123011 -0.513162828 -0.63409866
#> 41   0.84665868 -0.09928008  0.41154953 -1.255969351  0.33321213
#> 42  -1.31765844  2.11394675 -2.08095735  0.588625051 -0.47696181
#> 43  -0.48358964  0.50111128  0.09983433 -1.804422342  0.73031273
#> 44  -0.43953168  0.10381758 -0.28020942  0.144365024  0.53562473
#> 45   1.83016286  0.73259941  1.26070315  0.486978321  0.22187816
#> 46  -1.09320361  0.68453385  0.45420990 -1.046182311  1.15674763
#> 47   0.08610776  1.37670083  1.93545462  0.706010065 -0.50712068
#> 48   0.12654632  1.74157369 -0.59377930  1.243174898  0.53230598
#> 49   0.47148351  0.21512268  0.12746748 -2.439898420 -0.78164844
#> 50   0.02965696  0.67871022 -1.56781796 -0.286000790 -0.08631098
#> 51  -1.52316565  0.11614899  0.38251820 -0.006452206  0.06554403
#> 52  -1.39438349  0.22516714 -0.34849044  0.123102058 -0.01986720
#> 53   0.13957042  0.07911229 -0.45744898 -1.817791147 -1.63605058
#> 54  -0.53926360  0.75511518 -1.07808741 -1.062006683 -0.76763234
#> 55  -1.14927480  0.28034499 -1.28855286 -0.471859385  0.37523086
#> 56  -0.41484579 -0.03800856 -2.32039601  0.523311684  0.15415897
#> 57  -0.07908359  0.12880551  0.18444891  0.451773822  0.84791391
#> 58   1.68080612 -1.54225654  1.03443195  0.236923281  0.01784308
#> 59  -0.40722905  0.98762338  0.58590920 -0.412460316  0.17486934
#> 60  -0.63475402  0.70752455 -0.94374745 -0.098381485  0.40288008
#> 61   0.54733184 -0.93496033 -0.68081815  1.069952699  0.05493699
#> 62   0.12203797 -0.73240441 -1.32493764 -1.029568026 -1.92288107
#> 63  -0.74601022 -1.27012230  0.20698865 -0.828961157  0.38113816
#> 64  -1.06344964 -0.23864186  0.49987622 -0.982580735 -1.24205004
#> 65   1.29619939 -0.70436257 -0.31877881 -1.266796559  0.36320347
#> 66   0.30105628  0.55173045 -1.38419940 -1.388276075 -0.72595446
#> 67   1.02075443  0.52616616 -1.24422729  0.331470367  0.25042623
#> 68   0.26614796 -2.55166574 -0.33444885 -0.316804232 -1.43074930
#> 69  -1.13662116 -1.26365910  0.40319454  0.499383067 -0.76544314
#> 70  -0.29455223 -1.06323504  0.23668573 -0.964455249  0.50860678
#> 71  -0.65322109  1.17230555  1.35149693 -1.048744335 -1.18441174
#> 72  -1.43848898 -0.57290470 -1.90006886  0.264189912  0.69764199
#> 73   0.27195145 -0.01917456 -0.03038926  0.466321714 -0.35112136
#> 74  -0.28720544  1.77132796 -1.11486994  0.320730405  1.11018899
#> 75  -0.47416147 -0.44465567  0.28454024  0.211223063  0.70474469
#> 76   0.79018797 -1.50329998  1.04837466  0.506462673  1.65774799
#> 77   0.91598701 -1.11974665  0.54059712  0.319619641 -0.54258545
#> 78  -0.55252164  2.28525735 -1.39375240 -1.123588772 -0.42408263
#> 79  -0.10148556 -0.49133177 -0.59415171  0.112340356  0.62532182
#> 80   0.09554516  0.70110316  0.08473926 -0.918809298  0.71778866
#> 81   2.13914844 -0.68996966 -1.07774699 -0.224576429  0.61949298
#> 82  -0.01517728 -1.41440141 -0.82952129 -0.564072560  1.00325879
#> 83  -0.22599873  0.02951260 -1.18494646 -0.916005582  1.31431338
#> 84   0.21931385 -1.57159211 -1.32114399  0.677275951 -0.54847923
#> 85   0.58412407 -1.23883064 -0.97165957 -0.269622234 -0.64499125
#> 86   0.08442419 -0.18955332  1.48836470  1.340814502  1.39479068
#> 87   0.82989224  0.59662119 -0.60983856  0.097893198  1.67837922
#> 88   0.32110016  0.11071337 -0.49014649  1.027063590 -0.45837748
#> 89   0.17927000  1.15456977  0.03697649  0.526631040  0.26994114
#> 90   1.02073585 -0.50340874  0.38038923  1.283140661  0.42368033
#> 91   1.35968836 -1.43103912  1.24790457  1.239329052 -1.77880399
#> 92   0.13242927  0.56863755 -1.66649772  0.626112339  1.33060359
#> 93  -0.18114210  0.41654650 -0.48216094  1.762331377 -0.33937236
#> 94  -0.34175547  0.22978491 -1.03738363 -1.634959819  0.40207626
#> 95  -0.57551814 -1.43132167 -0.75905031  1.317503733  0.40839705
#> 96  -0.19959501 -1.02645811  0.44532282  0.863404867 -0.50664055
#> 97   0.32514801 -0.72754901  1.36717885 -0.142437963  1.37239652
#> 98  -0.28060398  0.71828049  0.21263582  0.400359737 -1.76520014
#> 99   0.02615514 -0.14522750  1.35787495  1.187707494  0.43830557
#> 100  0.07695597 -1.24578751  0.37828257 -0.712978446 -0.91380904