Access Outcomes

This is an getter method for outcomes in an S4 object, usually of class mediation_data.

Usage

outcomes(object)

Arguments

object

An object whose outcomes slot to modify.

Value

The output depends on the S4 class of the input object. If it is a multimedia model object, it will return a character vector of the outcome variable names. Otherwise it returns the outcome data.

Examples

exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
outcomes(exper)
#>               PHQ
#> 1    0.5558631832
#> 2   -1.2236690833
#> 3   -1.4542243198
#> 4    0.7482527416
#> 5    0.6911895591
#> 6    1.3021783167
#> 7   -1.1956295748
#> 8    1.0874834093
#> 9    0.0984541407
#> 10  -1.0002955490
#> 11   0.1550256939
#> 12  -1.2107053032
#> 13  -0.1736351974
#> 14  -0.2756448704
#> 15  -0.0361715338
#> 16   0.2038561489
#> 17  -0.8910370181
#> 18  -0.1630358095
#> 19  -1.7648930249
#> 20   0.2760004612
#> 21  -0.7090266056
#> 22   0.1216753133
#> 23  -0.2321011953
#> 24  -0.6869565329
#> 25  -0.8029961642
#> 26  -1.2752487683
#> 27   0.2729610820
#> 28   0.9177661885
#> 29  -0.3330136778
#> 30  -1.0240103507
#> 31   0.2149119090
#> 32   0.3163660796
#> 33  -2.3168323583
#> 34   0.8721301953
#> 35   0.9900721913
#> 36  -0.3619744317
#> 37  -0.3116603033
#> 38   0.6407370141
#> 39  -0.3041984762
#> 40   1.3114208953
#> 41   0.0285869350
#> 42   0.2988081686
#> 43   0.2098794500
#> 44  -1.7573061548
#> 45   0.9750318858
#> 46   0.3337227711
#> 47  -0.5425891768
#> 48   0.5934433993
#> 49   0.7548978345
#> 50  -0.7551830452
#> 51  -0.5620325981
#> 52  -0.4271940021
#> 53   0.3869809238
#> 54  -0.1476143229
#> 55  -0.8190309011
#> 56   0.1645761437
#> 57  -0.7056252438
#> 58  -0.9941529397
#> 59  -0.9397267435
#> 60  -1.5523849616
#> 61   0.7364420221
#> 62   1.7393890871
#> 63  -1.2783897734
#> 64   1.3432590773
#> 65  -0.7204703182
#> 66  -3.1724588793
#> 67  -1.6450771125
#> 68  -0.7294956200
#> 69   0.0004480106
#> 70  -0.3292138582
#> 71   0.1996555774
#> 72   0.8246933056
#> 73  -0.5115134118
#> 74  -2.7926952318
#> 75   0.1897717384
#> 76  -0.4306669818
#> 77  -0.9078029072
#> 78   0.4220484416
#> 79   1.9315923948
#> 80   1.1198550749
#> 81  -0.2683814394
#> 82  -0.6492696192
#> 83   1.4974887577
#> 84   1.1138030073
#> 85   1.6144237598
#> 86  -0.8796088871
#> 87  -0.3556191965
#> 88  -1.0693512100
#> 89  -0.7208885543
#> 90  -2.0129768144
#> 91   0.5002591626
#> 92   0.8834248512
#> 93  -0.9743459907
#> 94  -0.9078520012
#> 95   1.8898061342
#> 96   0.9085132033
#> 97  -0.1027397188
#> 98  -0.7500094741
#> 99  -1.5600735472
#> 100  0.1869838137