Outcomes Data in a Mediation Data Object

This is an accessor function to the @outcomes slot in a mediation data object. It returns the entire outcomes dataset, in contrast to outcomes() applied to a multimedia object, which only returns the names of the outcome variables.

Usage

# S4 method for class 'mediation_data'
outcomes(object)

Arguments

object

An object of class mediation_data.

Value

A data.frame whose rows are samples and columns different outcomes.

Examples

exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
outcomes(exper)
#>              PHQ
#> 1   -0.393536077
#> 2   -0.963333598
#> 3   -0.219115028
#> 4    0.374818220
#> 5   -0.086715491
#> 6   -0.201895831
#> 7   -0.097205239
#> 8    0.956991403
#> 9    0.746657546
#> 10   0.269307816
#> 11  -0.060140000
#> 12  -2.249867887
#> 13   1.093640942
#> 14   0.534374533
#> 15   0.395957872
#> 16  -0.209636205
#> 17   1.708589095
#> 18  -0.790794093
#> 19  -0.568207619
#> 20   1.001387949
#> 21  -0.458125800
#> 22  -0.424620114
#> 23   0.422604895
#> 24   0.405273841
#> 25  -0.779336725
#> 26   0.877076729
#> 27  -0.154420649
#> 28   0.910117065
#> 29   0.023111809
#> 30  -0.234114111
#> 31  -0.016797210
#> 32   0.668456216
#> 33  -0.229797500
#> 34   0.234556756
#> 35  -0.667952121
#> 36   0.276662541
#> 37   1.327078819
#> 38  -0.300154623
#> 39  -1.442420234
#> 40  -1.071384954
#> 41   0.932559165
#> 42   0.008695232
#> 43   0.641263260
#> 44   0.946711328
#> 45  -0.018085638
#> 46   0.303651140
#> 47   1.276867048
#> 48   2.021410379
#> 49  -1.305214886
#> 50   0.961518664
#> 51  -0.028351671
#> 52  -0.163106109
#> 53   0.496603497
#> 54   0.086050648
#> 55   1.924381769
#> 56   0.010375252
#> 57   0.274514256
#> 58  -0.600876274
#> 59  -0.174487420
#> 60  -2.022422197
#> 61   0.448402172
#> 62  -0.435232974
#> 63   0.599709066
#> 64   0.504464056
#> 65   0.051946457
#> 66  -0.400894247
#> 67   1.290208973
#> 68  -0.036818563
#> 69   1.810922927
#> 70  -1.941260748
#> 71   0.778363744
#> 72  -0.555691891
#> 73  -1.883298553
#> 74  -1.437915093
#> 75  -0.789960568
#> 76  -0.229661176
#> 77  -1.414493245
#> 78  -0.406737751
#> 79   0.474896695
#> 80   0.508970065
#> 81   1.309435110
#> 82   0.775685377
#> 83   1.244277308
#> 84  -0.670778177
#> 85   1.027167800
#> 86   0.110047652
#> 87   0.169136797
#> 88  -0.487560128
#> 89   0.375037921
#> 90  -0.872785097
#> 91  -1.763655549
#> 92  -0.131420498
#> 93   0.907619152
#> 94  -2.323299375
#> 95  -2.263386249
#> 96   0.655683860
#> 97  -0.527277818
#> 98   0.258998995
#> 99  -1.191954030
#> 100  1.427593741