Sample from a Random Forest Model

This assumes a continuous response, so that the out-of-sample MSE can be used to estimate the outcome variability \(\sigma\).

Usage

rf_sampler(fits, newdata = NULL, indices = NULL, ...)

Arguments

fits

The fitted RF model from which to draw samples.

newdata

A data.frame containing new inputs from which to sample responses. If NULL, defaults to the data used to estimate fit.

indices

The coordinates of the response from which to draw samples.

...

Additional parameters passed to rf_model's predict method.

Value

y_star A data.frame of samples y associated with the new inputs.

Examples

m <- rf_model()
fit <- estimator(m)(mpg ~ hp + wt, data = mtcars)
rf_sampler(fit, mtcars)
#> # A tibble: 32 × 1
#>      mpg
#>    <dbl>
#>  1  11.4
#>  2  15.0
#>  3  15.9
#>  4  23.2
#>  5  27.6
#>  6  33.0
#>  7  16.7
#>  8  33.2
#>  9  13.6
#> 10  30.8
#> # ℹ 22 more rows

prf <- parallelize(ranger::ranger)
fit <- prf(mpg + disp ~ hp + wt, data = mtcars)
rf_sampler(fit, mtcars)
#> # A tibble: 32 × 2
#>      mpg   disp
#>    <dbl>  <dbl>
#>  1  22.5   926.
#>  2  27.1 -2008.
#>  3  26.7   602.
#>  4  25.6 -3524.
#>  5  21.9  1535.
#>  6  25.5  -824.
#>  7  10.6  1407.
#>  8  24.8  -567.
#>  9  17.0 -2236.
#> 10  23.4  4554.
#> # ℹ 22 more rows