Treatments in a Mediation Data Object

This is an accessor function to the @treatments slot in a mediation data object. It returns the entire set of observed treatments, in contrast to treatments() applied to a multimedia object, which only returns the names of the treatment variables.

# S4 method for class 'mediation_data'
treatments(object)

Arguments

object

An object of class mediation_data.

Value

A data.frame whose rows are samples and columns different treatments.

Examples

exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
treatments(exper)
#>     treatment
#> 1     Control
#> 2   Treatment
#> 3     Control
#> 4   Treatment
#> 5     Control
#> 6   Treatment
#> 7   Treatment
#> 8     Control
#> 9   Treatment
#> 10  Treatment
#> 11  Treatment
#> 12    Control
#> 13    Control
#> 14    Control
#> 15    Control
#> 16  Treatment
#> 17    Control
#> 18    Control
#> 19  Treatment
#> 20    Control
#> 21  Treatment
#> 22    Control
#> 23    Control
#> 24  Treatment
#> 25  Treatment
#> 26    Control
#> 27    Control
#> 28    Control
#> 29    Control
#> 30  Treatment
#> 31    Control
#> 32    Control
#> 33    Control
#> 34    Control
#> 35    Control
#> 36    Control
#> 37    Control
#> 38    Control
#> 39  Treatment
#> 40    Control
#> 41  Treatment
#> 42    Control
#> 43  Treatment
#> 44    Control
#> 45    Control
#> 46    Control
#> 47    Control
#> 48    Control
#> 49    Control
#> 50    Control
#> 51    Control
#> 52    Control
#> 53    Control
#> 54  Treatment
#> 55  Treatment
#> 56    Control
#> 57  Treatment
#> 58  Treatment
#> 59  Treatment
#> 60    Control
#> 61  Treatment
#> 62  Treatment
#> 63  Treatment
#> 64    Control
#> 65  Treatment
#> 66  Treatment
#> 67  Treatment
#> 68  Treatment
#> 69  Treatment
#> 70  Treatment
#> 71  Treatment
#> 72    Control
#> 73  Treatment
#> 74  Treatment
#> 75    Control
#> 76    Control
#> 77    Control
#> 78    Control
#> 79    Control
#> 80  Treatment
#> 81  Treatment
#> 82  Treatment
#> 83    Control
#> 84  Treatment
#> 85  Treatment
#> 86  Treatment
#> 87  Treatment
#> 88    Control
#> 89    Control
#> 90  Treatment
#> 91  Treatment
#> 92    Control
#> 93  Treatment
#> 94  Treatment
#> 95  Treatment
#> 96  Treatment
#> 97  Treatment
#> 98  Treatment
#> 99    Control
#> 100   Control