A Demo Dataset (Spline)

This is a simple dataset with nonlinear relationships between the outcome and mediators. It is used simply to illustrate the design of multimedia. The mediator->outcome effect is generated from a random spline function.

demo_spline(n_samples = 5000, tau = c(2, 2))

Arguments

n_samples

The number of samples to generate in the toy example

tau

The true direct efefcts associated with the two outcomes. Defaults to 2, 2.

Value

xy A data.frame whose columns include the treatment, mediation, and outcome variables.

Examples

demo_spline()
#> # A tibble: 5,000 × 4
#>    outcome_1 outcome_2 mediator treatment
#>        <dbl>     <dbl>    <dbl> <fct>    
#>  1    1.21     -0.0536   -1.20  0        
#>  2    1.22     -1.12     -4.01  0        
#>  3    0.418    -0.658    -1.50  0        
#>  4    0.0883   -1.17     -1.92  0        
#>  5   -0.444     1.18      1.57  0        
#>  6   -0.517    -0.212    -0.772 0        
#>  7   -0.122     1.10      0.525 0        
#>  8    0.839    -1.34     -3.63  0        
#>  9    1.11     -1.21     -2.74  0        
#> 10   -1.01      0.961     1.07  0        
#> # ℹ 4,990 more rows
demo_spline(20)
#> # A tibble: 20 × 4
#>    outcome_1 outcome_2 mediator treatment
#>        <dbl>     <dbl>    <dbl> <fct>    
#>  1    0.912      0.850  -1.98   0        
#>  2    0.145     -1.01   -5.25   0        
#>  3    1.37       1.43   -1.76   0        
#>  4    0.790      0.737  -1.26   0        
#>  5   -0.0155    -1.12   -5.47   0        
#>  6    0.978      0.936  -1.95   0        
#>  7   -0.506     -0.461   0.0663 0        
#>  8    1.11       0.922  -2.40   0        
#>  9    0.713     -0.280  -5.37   0        
#> 10   -0.0722    -0.225  -0.395  0        
#> 11    0.565      0.886   0.472  1        
#> 12    2.41       2.55   -0.0505 1        
#> 13    1.80       0.399  -4.21   1        
#> 14    1.15       3.04    4.53   1        
#> 15    1.57       1.61    0.0493 1        
#> 16   -0.320      1.68    1.71   1        
#> 17    1.30       2.78    4.80   1        
#> 18    3.40       0.664   6.89   1        
#> 19    0.712      3.59    3.57   1        
#> 20    1.99       1.01   -3.27   1        
demo_spline(20, c(0, 10))
#> # A tibble: 20 × 4
#>    outcome_1 outcome_2 mediator treatment
#>        <dbl>     <dbl>    <dbl> <fct>    
#>  1    0.621     -1.25     2.78  0        
#>  2    0.269      1.27    -4.94  0        
#>  3   -0.339      0.313   -0.665 0        
#>  4   -0.558      0.362   -3.05  0        
#>  5    0.443      0.955   -2.81  0        
#>  6   -0.138      0.529   -0.947 0        
#>  7   -0.236      0.556   -3.03  0        
#>  8    0.243      0.855   -3.20  0        
#>  9   -0.626      1.25    -6.07  0        
#> 10   -3.42       0.683   -7.93  0        
#> 11    0.943      9.06     2.89  1        
#> 12    0.680     11.1      6.12  1        
#> 13    1.32       9.17     2.77  1        
#> 14   -0.0837     9.60     3.86  1        
#> 15    0.239      9.95     0.226 1        
#> 16    0.375     10.5     -0.262 1        
#> 17    0.309      8.37     2.56  1        
#> 18    0.964      8.62     1.79  1        
#> 19   -1.16       9.98    -2.19  1        
#> 20    0.150      8.10     2.25  1