A Demo Dataset (Spline)

This is a simple dataset with nonlinear relationships between the outcome and mediators. It is used simply to illustrate the design of multimedia. The mediator->outcome effect is generated from a random spline function.

Usage

demo_spline(n_samples = 5000, tau = c(2, 2))

Arguments

n_samples

The number of samples to generate in the toy example

tau

The true direct effects associated with the two outcomes. Defaults to 2, 2.

Value

xy A data.frame whose columns include the treatment, mediation, and outcome variables.

Examples

demo_spline()
#> # A tibble: 5,000 × 4
#>    outcome_1 outcome_2 mediator treatment
#>        <dbl>     <dbl>    <dbl> <fct>    
#>  1    1.16     -0.737    -1.25  0        
#>  2    0.0852    0.962     1.27  0        
#>  3    0.970     1.10      0.737 0        
#>  4   -0.277    -1.41     -2.80  0        
#>  5   -1.87     -0.596    -4.59  0        
#>  6    0.807    -0.632    -0.771 0        
#>  7    0.280    -1.52     -1.56  0        
#>  8    1.55     -0.201    -0.896 0        
#>  9    1.03      0.0702   -0.241 0        
#> 10    1.39     -0.0860   -0.646 0        
#> # ℹ 4,990 more rows