Convert a Summarized Experiment to a data.frame

This is a small helepr function to concatenate the colData and assays within SummarizedExperiment objects into a single data.frame. This is not necessary for the essential multimedia workflow, it is only exported for potential independent interest.

Usage

exper_df(exper)

Arguments

exper

An object of class SummarizedExperiment.

Value

A data.frame combining all slots of a multimedia object.

Examples

demo_joy() |>
    multimedia:::exper_df()
#>            ASV1         ASV2        ASV3         ASV4        ASV5 treatment
#> 1    1.58627044 -1.270817406  0.47355816  1.327535467  1.79973279 Treatment
#> 2    1.28349033  0.761429685 -0.35115503  0.984848460  0.87838002   Control
#> 3   -1.63767160 -0.815159761 -1.70636296 -0.813535755  0.28586006 Treatment
#> 4    0.38659127 -0.128506917 -0.80867258  0.473594121 -0.95967703 Treatment
#> 5   -0.08902674 -0.823100202  0.28510448 -0.515760956 -0.53320223 Treatment
#> 6   -0.99350756  1.564026860  0.43230044  0.250033410 -0.11782507 Treatment
#> 7    1.17215023  0.841391250  0.14364210  0.003104126  0.07452745   Control
#> 8    1.05572949  0.438907366  0.52763928 -1.823989181 -1.00798659   Control
#> 9   -0.32600601  0.246940098  0.44650789  2.129979562  0.37478872   Control
#> 10  -0.72647015  1.233203776  0.77658150 -0.123315613 -0.06872677 Treatment
#> 11   0.34169810  0.630828215 -0.41345500  1.578676459  0.23974848 Treatment
#> 12   3.20037351  0.197140651 -0.02505539 -0.606482258 -0.58121283   Control
#> 13  -0.47908455  0.003481963 -1.24961311 -0.291555552 -1.29134746   Control
#> 14  -0.18288321 -0.427015963 -1.28348502 -0.102881164 -0.53327888 Treatment
#> 15   1.44479529  1.041334511 -1.11083465 -1.310438405  0.64692851   Control
#> 16   0.13965423  0.496525577  1.48499723  0.394953291 -0.35640734 Treatment
#> 17  -0.96153114 -1.503952134 -0.72174259 -0.566449400  0.28860802   Control
#> 18  -0.07926306  0.061039991 -0.27610795 -0.809003433  0.92714656 Treatment
#> 19   0.85756637  1.474323275 -1.60541585 -0.908704508  1.17461514 Treatment
#> 20  -0.58878033 -1.716235564  0.89613992 -1.301569176 -0.77677970 Treatment
#> 21  -0.18012811 -0.609946818 -1.03453798  0.023988134  2.55954859   Control
#> 22  -0.25844751  1.342987895  0.22873353  1.961102697 -0.77708530 Treatment
#> 23   0.26614093  0.260096642 -0.53788427  1.275755935 -0.40450943 Treatment
#> 24  -0.46514597 -2.053895381 -0.32612414 -0.260130375  2.06092649 Treatment
#> 25   0.23429498  1.622295535  1.00743659  0.100491930  1.28422955 Treatment
#> 26   1.14286280 -0.148631739 -0.72305676 -1.873392649 -1.67470841   Control
#> 27   0.25657478  0.371072535  0.57550366 -1.012667710  0.40480578   Control
#> 28  -0.56846267 -1.825340724 -0.42482482 -0.089035389  0.92610525   Control
#> 29  -1.32554289 -0.991590179  0.03298245  0.053868831  0.65683967   Control
#> 30   1.39155927  1.057621448 -0.79703199  0.609737063  0.57115505   Control
#> 31   0.01525205  0.094236024 -0.49289083 -1.261132877  0.18156946   Control
#> 32   1.27465544  0.270923702 -0.27707476 -0.991021696 -0.81040912 Treatment
#> 33  -1.22475870 -0.545388243 -1.05719474  0.999114419  0.59899180   Control
#> 34  -0.59741072  0.814929891  1.45096107  0.878623883  0.55535940   Control
#> 35  -0.78821419 -0.724109628  0.39081498 -0.184239346 -0.25235385 Treatment
#> 36  -0.26021156 -0.241461971  1.62661986  1.099360735 -1.02404364   Control
#> 37   0.57254127  0.405081653  1.77134995 -1.097797890 -0.35800190   Control
#> 38   1.78350186 -1.414076457 -0.31743255  1.816345051 -0.88363996   Control
#> 39  -0.81399342  2.721289200 -0.88150311 -1.026512595 -0.13707001   Control
#> 40  -0.86838061 -0.087329626  0.62222831 -0.878041415 -0.84880761 Treatment
#> 41   0.11714161 -0.312618377 -0.04920851  0.540951861 -0.27811565   Control
#> 42   0.73614880 -1.176523541 -0.04363507  0.942136464  0.07201662   Control
#> 43   0.91375719  0.238012244  0.60731814 -0.173614772 -0.53169246 Treatment
#> 44   0.89117857  0.090764770  1.23900707 -0.532552322 -2.05595123 Treatment
#> 45   1.00747204 -0.246444630 -0.85785520  0.982932055  0.14935632 Treatment
#> 46  -0.74106400  0.625220994 -0.20881098 -1.275913579 -0.17877248 Treatment
#> 47  -1.55524653 -0.690930198  0.11460587 -0.838653371  0.04262537 Treatment
#> 48  -0.62627030 -1.829384330  1.37232090  0.730008088 -1.95053599 Treatment
#> 49   1.84637580  0.423529308 -0.67761246 -1.621729427 -1.56972802   Control
#> 50   1.07385076 -1.088342396  0.46019799 -0.635834553  0.95503040   Control
#> 51  -1.16579413  0.191253246  0.64259303 -0.576783250  0.14359135   Control
#> 52   0.24878478 -0.338721198  1.45657602  1.564919468  0.76650991   Control
#> 53   0.56004605 -0.640636982 -1.14380130 -0.819954837 -1.81153746 Treatment
#> 54  -0.41780805 -0.057467649  0.74174598  1.155938228  1.65715897   Control
#> 55   0.46396176  1.214572695  0.77151715  0.321918845  1.42533964   Control
#> 56  -0.82186060 -1.486282181  0.01277595  0.994436823 -0.60260327 Treatment
#> 57  -0.62406229 -0.203005695  0.48547394 -1.038744347  0.46509781   Control
#> 58  -0.43097227  0.468821424  0.16988783  0.484128972 -0.12100265 Treatment
#> 59  -0.89466207  1.103755507 -1.43863176  1.141097816  1.21956532 Treatment
#> 60  -0.14969970 -1.293072985 -0.16101323  0.525970228  1.12312429   Control
#> 61  -1.01909005  1.711385008 -1.95112973 -0.374341296 -1.64651431   Control
#> 62  -0.58274406  0.294569542  1.69843862 -1.546100545  0.61987727 Treatment
#> 63  -0.03571180  0.920435625 -0.49912072 -1.149969485  0.01408439   Control
#> 64   1.40177347 -0.025735528 -0.11346245  0.862008741  1.44498682   Control
#> 65   1.13226630 -0.303936367 -0.40564506 -1.334668801 -1.59226112   Control
#> 66  -0.02966530  0.133594424  0.46875839 -0.956704768 -2.65780141   Control
#> 67   0.15807338 -0.192656038 -0.88337776 -1.185885804  1.06540822 Treatment
#> 68  -0.61944846  1.938239012 -2.45148418 -2.769338872  1.87008787   Control
#> 69   2.99929388 -0.976372938 -0.58375115  1.356586827 -0.38017750   Control
#> 70  -0.04122614 -0.419727608 -0.61828358 -0.592086341  0.49071085 Treatment
#> 71  -1.33316626 -0.592201172  3.42674258  0.089999472 -0.16844765 Treatment
#> 72  -0.50715060  1.673334734  0.49002428  0.262921656  0.68147548   Control
#> 73  -2.14227645  0.199389603  0.61804584  0.304980353  2.04932083 Treatment
#> 74  -0.55923328 -1.136982195 -1.32228144  0.668450125  0.91024186 Treatment
#> 75  -0.02634214 -1.686084427  0.20796774 -1.981844580 -2.28354716 Treatment
#> 76   0.30399563 -0.017745199  0.63024562  0.015938549 -2.13619996 Treatment
#> 77   0.20119668  0.234489858  0.20479962 -0.292811578 -1.60453669   Control
#> 78  -0.24700111 -0.254676403 -0.23331737 -0.143302247  1.06285528   Control
#> 79   0.69357722 -0.556095052 -1.83480022  0.768775152 -0.57293102   Control
#> 80   0.77005440  0.929259334 -0.88504769 -0.473068020  1.04371956 Treatment
#> 81   1.06084789 -0.277567884  0.28352055 -0.864912652 -0.59316130 Treatment
#> 82   0.27456197 -1.864619743 -1.68795039  0.536056331 -0.26378038   Control
#> 83   0.99322473 -0.614199038  2.43132639  0.035478971  0.71277321 Treatment
#> 84  -1.40138367  0.378225162 -0.50878129  0.646726518  0.47951261   Control
#> 85   0.38034104 -0.333289407 -1.02984293 -0.792201733 -0.59800241 Treatment
#> 86  -1.36735152  0.526560354 -0.50551831 -0.564946883  1.68183516   Control
#> 87   0.08520366 -1.645498754 -0.12167841  0.789811755 -1.08016173   Control
#> 88   0.39312803 -1.065956301  0.06972973 -0.144827328  0.02877809 Treatment
#> 89  -0.20177851 -0.051609860 -0.70871468  1.675717956 -1.12526010 Treatment
#> 90  -0.93637697  1.583032449 -0.34952047  0.049161504  0.18624634 Treatment
#> 91   0.05480242 -1.782828116 -0.62765630 -1.780051841 -0.07655138 Treatment
#> 92   0.69237906 -0.075494831  0.06798136  0.772521577  0.52686928   Control
#> 93   1.54776356 -0.137296038 -0.72919677  1.285729325  1.44779132   Control
#> 94   0.11028658  0.892853936  1.04501638 -0.947769052 -0.43073365   Control
#> 95  -0.29930994 -0.806580090 -0.03393759 -0.918597847  1.17554749 Treatment
#> 96  -1.32532641  0.817378151  0.51153836  0.539991553  1.03468629 Treatment
#> 97  -0.31405632 -0.678083464 -0.99001541 -0.086217641  0.10173840   Control
#> 98   0.38246219  0.122375602  0.26273720 -1.045780469 -0.53128803   Control
#> 99   0.13227339 -0.048221392  0.04871549  0.304905845 -1.00762735 Treatment
#> 100  0.68256317 -1.173197679 -1.95319544 -0.722385532  0.30709804 Treatment
#>              PHQ
#> 1    0.020155699
#> 2    0.923062823
#> 3    0.296590973
#> 4    0.118353284
#> 5   -0.790453105
#> 6    0.617933574
#> 7    0.105041589
#> 8    0.917043200
#> 9   -0.712000770
#> 10   0.998885713
#> 11   0.506492645
#> 12  -1.691070102
#> 13   0.413496665
#> 14   1.448829376
#> 15   0.708824176
#> 16  -0.766818550
#> 17  -1.191232224
#> 18  -0.081658464
#> 19   0.611778262
#> 20  -0.215080181
#> 21  -0.844169651
#> 22  -2.080398111
#> 23   0.310214596
#> 24   0.382170502
#> 25   0.522906118
#> 26   0.398002321
#> 27  -0.699682678
#> 28   1.694332545
#> 29  -2.396042158
#> 30  -0.958826197
#> 31  -0.778922069
#> 32  -1.237716153
#> 33   1.373720912
#> 34  -0.551548421
#> 35  -0.603197603
#> 36   0.003078969
#> 37   0.382920153
#> 38   0.351893181
#> 39   0.333326813
#> 40  -0.434050122
#> 41   1.328610117
#> 42   0.771236970
#> 43  -0.008743135
#> 44   0.929032523
#> 45  -0.705050154
#> 46  -1.635936370
#> 47  -0.481898254
#> 48   0.289442554
#> 49   0.108698008
#> 50   2.669657377
#> 51   0.081576722
#> 52  -0.425701991
#> 53   0.181395765
#> 54  -0.761303139
#> 55  -1.148546091
#> 56  -1.477080744
#> 57  -0.678914251
#> 58   0.542552640
#> 59   0.571660999
#> 60  -0.936703146
#> 61   1.000848248
#> 62   0.242451632
#> 63   0.864212354
#> 64   0.755961315
#> 65   1.536701780
#> 66   0.834803633
#> 67  -0.833979191
#> 68   1.255173122
#> 69  -0.255151028
#> 70   0.495407083
#> 71  -0.624752466
#> 72   0.920721056
#> 73  -1.165755913
#> 74  -0.740555680
#> 75   2.026509780
#> 76   0.879698832
#> 77   2.865341242
#> 78  -0.141284487
#> 79   0.274117235
#> 80  -0.738428242
#> 81   1.455993151
#> 82   0.763539212
#> 83   2.031081867
#> 84  -1.781681939
#> 85   0.044634811
#> 86   0.386299674
#> 87   0.319184988
#> 88   1.303122934
#> 89  -0.734209133
#> 90  -2.646414799
#> 91  -2.243226467
#> 92   0.270122113
#> 93   0.061001403
#> 94  -0.655419989
#> 95  -1.062160071
#> 96  -0.196648063
#> 97   0.552798980
#> 98   0.085138698
#> 99   0.786797265
#> 100  0.376360151