Convert a Summarized Experiment to a data.frame

This is a small helepr function to concatenate the colData and assays within SummarizedExperiment objects into a single data.frame. This is not necessary for the essential multimedia workflow, it is only exported for potential independent interest.

exper_df(exper)

Arguments

exper

An object of class SummarizedExperiment.

Value

A data.frame combining all slots of a multimedia object.

Examples

demo_joy() |>
    multimedia:::exper_df()
#>             ASV1        ASV2         ASV3        ASV4        ASV5 treatment
#> 1   -0.782300340 -0.70354655 -3.085106383 -1.44937373 -1.52974524   Control
#> 2    0.346099273 -0.42673424 -0.510881446 -0.09827899  0.72633032 Treatment
#> 3    0.514091951  0.22179294 -0.898263579 -1.28313630 -0.75994608   Control
#> 4    1.294934858 -0.28187187  0.621217092  1.86880227  0.22872538 Treatment
#> 5    0.352033430  0.47733485  0.875454111  0.32916481  1.00731798   Control
#> 6   -0.193778564 -0.32783551 -2.373395115  1.56731606  0.04860601   Control
#> 7    0.607246790 -0.13899514 -0.795431378  1.07392592  1.30196556   Control
#> 8    3.466623216  0.63553345  0.656342159  0.51018702  0.77331618 Treatment
#> 9    1.273467606  1.66836733  0.824550250  0.99743274 -0.81867298   Control
#> 10  -0.004955086  1.63485523  0.485218810 -1.85386889 -1.07237704 Treatment
#> 11   0.006056590  0.07829037  2.103318842 -0.53223532  2.00447177 Treatment
#> 12   0.582897838 -0.01875546  1.111052555  1.13681414 -0.18596497 Treatment
#> 13  -0.636339068  2.15133055 -0.292651031 -1.30901057  0.87178063   Control
#> 14   0.478329441  0.43866808  1.700390189  0.76734819 -0.45950441 Treatment
#> 15  -0.516066636 -0.05003374  1.345242273 -0.19005907 -0.21208755   Control
#> 16   1.173761180 -0.05449931  0.378617015 -0.45440956  0.36941274 Treatment
#> 17   0.527858040  3.20510178  0.262430240 -0.76157867 -0.58686809   Control
#> 18  -0.401446329 -1.02205860  0.057858674 -0.05757483 -1.67889346 Treatment
#> 19   0.274497455  0.85638283  0.207409579  1.12636844  0.85500514   Control
#> 20   0.995934209  0.42343857 -1.974090209 -2.14070644 -0.22275038 Treatment
#> 21  -0.313310478  0.40298481 -0.476208165  0.40302875  0.52022235   Control
#> 22  -1.542884914 -0.26695760  2.285765750 -0.24837436 -0.07547115 Treatment
#> 23  -1.347816931  0.26305285  0.378541634 -0.08198606 -0.86758445   Control
#> 24   0.633476808  0.41447151  0.562066741  1.56030977  1.49757385 Treatment
#> 25   0.306755934 -2.30344943 -0.348294499  2.06934897  0.08444860   Control
#> 26   0.427162730 -0.67177219 -1.337173324 -1.50366588  0.56296419   Control
#> 27  -0.479869169  0.35598324  0.613131321  1.67374151  0.43247814   Control
#> 28  -0.162113769  0.02212038  0.105962275 -0.79322470 -1.25038358 Treatment
#> 29  -1.023374716  0.03041205 -0.493593682  1.46638923  0.49454434 Treatment
#> 30  -0.353598613  0.71360127 -0.354791571 -0.47055903  0.33358174   Control
#> 31   0.295707379 -0.19311208  1.201445121 -0.16902562  0.48897308   Control
#> 32  -0.843026548 -2.21193787 -0.391578370  0.03965610  0.24941897 Treatment
#> 33   0.093150733  0.60676330  2.028937545  0.30069697  0.41564740 Treatment
#> 34   0.984684044  0.39717018  0.225414140  1.24733571  0.89417881   Control
#> 35  -0.364512384 -0.45068665 -0.685705095 -0.17850987  1.12014596   Control
#> 36  -0.520444679  0.24183793  0.590976582  0.41573286 -0.65466246   Control
#> 37  -0.533331137 -1.05744032  1.395766112 -0.14134510  1.31875754   Control
#> 38   0.415137951  0.67010362  0.398505827  1.00070591 -2.76153430   Control
#> 39   0.264808810 -0.62214809 -1.211486717 -1.22813948 -0.42720817 Treatment
#> 40  -0.707789745  0.46488138  1.164699584 -0.51983487 -0.02305103   Control
#> 41  -0.141147068  0.01539255  2.040716516  1.69004187 -0.72625836   Control
#> 42   0.091252279 -0.46594253 -0.036912249  0.98835007 -0.31992403   Control
#> 43   0.178235796 -0.16381636  1.762587571 -2.37449590 -2.05156301   Control
#> 44  -0.205118355  0.80462277  0.421248556  0.52631795  0.87272644   Control
#> 45   0.662124829 -0.51641927 -1.281441982 -0.26444871 -0.77998588 Treatment
#> 46   1.776724741 -0.11719815  0.411353110  0.24257916  0.07834463 Treatment
#> 47  -0.690619048 -0.76864496 -0.052624675 -2.20928737 -0.16657743   Control
#> 48  -0.081422180 -1.24542847 -0.939899582 -1.26833011  2.14293831 Treatment
#> 49  -0.296774476 -1.05416089  1.628534563 -0.93077574 -0.23661757   Control
#> 50   0.373486180  0.50922061  0.577730245 -0.40850266 -1.39076618 Treatment
#> 51   0.303544840  0.21428588 -0.612975826  0.37280594 -0.24269561 Treatment
#> 52  -0.852631912 -0.23219310  0.305823924  0.22203487 -0.40897900   Control
#> 53   1.013231616  0.45901711 -0.867105614  0.79194643  0.88498136   Control
#> 54   0.205271080 -1.69940043  1.495816349 -0.55571103  1.56513642   Control
#> 55   0.086058337 -1.27777491  1.223841360 -0.90037098  2.21428833   Control
#> 56   0.298300979  0.08074379  1.860448589  0.84650469  0.85445754 Treatment
#> 57   0.683394336 -0.50751044  0.751456290 -1.29363670 -1.01821493 Treatment
#> 58  -0.406321263 -0.34294575 -0.192868797 -0.70363746 -0.22934284   Control
#> 59  -1.327429784 -0.65601011  0.469665643 -0.15336402 -0.34517800 Treatment
#> 60   2.453563483  0.47405423 -1.172875358  0.60343207  1.60849338 Treatment
#> 61   0.485875119  0.44523536  0.543830476 -0.77177410  0.26157399 Treatment
#> 62  -0.981058873 -1.39776218 -0.545878902  0.21148932 -1.27024515 Treatment
#> 63   0.130752242 -0.91465294 -1.610224926  0.97291031 -0.55112111 Treatment
#> 64   0.074955899  0.50692240  0.272584491 -1.38593058 -0.73541835   Control
#> 65  -0.041629565  0.23988105  0.369799166 -0.24910014 -0.26153482   Control
#> 66   1.604351654 -0.65826009  0.002241616 -0.19877798 -2.18890064   Control
#> 67  -1.097415834  0.94544902  0.898573804 -0.72061567 -1.18371423   Control
#> 68   1.903525066 -1.73906475 -0.967294107  0.48850682  0.85617000 Treatment
#> 69  -0.603538425 -0.82466634  1.523033813 -0.38760253  0.18392778   Control
#> 70   0.207640199 -0.77677437 -0.267872591  0.13040933  1.40424251 Treatment
#> 71   0.496129384  0.15127714  0.485922991 -0.13044758  1.05327737   Control
#> 72   0.279517030 -0.58079839  1.386581521  0.40733757  0.38640116 Treatment
#> 73  -0.890410034  1.29487534  0.971109233 -1.50604405 -0.61653285 Treatment
#> 74  -0.380797116  0.58947418  0.463520040  1.66029894 -0.26195976 Treatment
#> 75  -0.507452437  0.15649728  1.418706581 -0.17775041  0.93145878   Control
#> 76  -0.032311881  0.23064099  1.084425701 -0.03511277 -1.26143998   Control
#> 77  -0.227546033  0.82011314  0.914472782 -2.48203170  0.49516382   Control
#> 78   2.197867955 -0.40605616 -1.058554096 -0.75233987  0.98482654   Control
#> 79  -0.203465299 -0.15661689  1.926114935  3.72819336 -0.16385131 Treatment
#> 80  -1.195466581 -0.67219047  0.993603321 -1.23620775 -1.36009194   Control
#> 81  -1.517288375  1.15311771  0.515174750  0.82932682 -2.12647230 Treatment
#> 82  -1.319363200 -0.53228374  0.105384256 -1.22734818 -1.29497230   Control
#> 83   1.129840216  0.97384870  0.624863046 -0.03216184 -0.85984875   Control
#> 84  -0.784910874 -1.32793580 -0.668362245 -0.85720107 -1.41075454   Control
#> 85  -0.517027514  1.84863892  0.364498800  0.91553548  0.58480961 Treatment
#> 86   0.428344760  0.70954363  0.010621504 -1.40610457 -0.28228025   Control
#> 87   0.074517515 -0.76521283 -1.282552637 -0.78439910  0.06652903   Control
#> 88   0.872639981  1.20908087 -0.817899775  0.65363786  1.18726369 Treatment
#> 89   1.452136998  1.52638532 -0.404115848  1.71591748 -1.02427032   Control
#> 90   0.588504625 -1.11423377  1.422491414  0.20653781  0.17102495 Treatment
#> 91  -1.212458976 -0.83046388  0.449407306 -0.18161381 -0.40014684 Treatment
#> 92   2.396426815  0.57079049 -1.154529142 -0.15250654 -0.92544032   Control
#> 93   2.251591546 -0.08237616  0.614812510 -0.58763104 -0.99555523   Control
#> 94   0.030548373  0.36628803 -1.208204528  1.25856824 -0.52436257   Control
#> 95   0.662044048  0.09913842  0.806423530  1.03042899 -2.33683999   Control
#> 96  -1.114478736  1.16957945 -0.374628415  0.84300965 -0.19215890 Treatment
#> 97  -0.508279324 -0.84436086 -0.152780750 -1.02173563  0.36671791   Control
#> 98  -0.520389451 -1.61628985  0.946631754  0.96935807  0.35411357 Treatment
#> 99  -1.316297017  1.37975595  1.510029521  0.15941176  0.58963938 Treatment
#> 100 -0.383776358 -0.32265238  0.255736449 -0.48341939 -0.76912129 Treatment
#>              PHQ
#> 1   -0.346559362
#> 2    0.264621344
#> 3    2.018936646
#> 4   -0.435984089
#> 5   -1.969932889
#> 6    0.973929648
#> 7    1.110937865
#> 8    0.008575746
#> 9   -1.102066036
#> 10  -0.855993954
#> 11   1.492408055
#> 12  -0.014259710
#> 13   0.723322785
#> 14  -1.060102368
#> 15  -1.346010919
#> 16  -0.188701431
#> 17   0.336566930
#> 18   0.194137289
#> 19   0.292637400
#> 20  -1.551280969
#> 21  -0.063192675
#> 22   1.225293990
#> 23  -0.167231457
#> 24  -0.086690825
#> 25  -0.594402805
#> 26  -0.769676516
#> 27   0.194991782
#> 28   1.359153537
#> 29   1.814908214
#> 30   1.753847336
#> 31   0.308994795
#> 32  -1.131582762
#> 33  -1.684655986
#> 34   0.599808656
#> 35  -0.570400118
#> 36   0.109350519
#> 37   0.089564911
#> 38   0.449856262
#> 39   0.478608881
#> 40   1.339174630
#> 41   1.680668886
#> 42  -0.625473861
#> 43   1.219626479
#> 44   0.304623374
#> 45   0.114719764
#> 46   0.498836253
#> 47  -1.462792278
#> 48   0.611001172
#> 49   0.086839044
#> 50  -0.318564998
#> 51  -0.277582496
#> 52  -0.104985286
#> 53  -0.388651062
#> 54   0.019814571
#> 55   0.723515364
#> 56   0.871626888
#> 57   0.122993181
#> 58  -0.027201659
#> 59   0.948923725
#> 60  -0.002816949
#> 61   0.260059036
#> 62   1.049104154
#> 63  -1.484457581
#> 64   0.049979811
#> 65  -1.051595718
#> 66   0.875766624
#> 67  -1.278854639
#> 68  -0.324864363
#> 69  -0.648554652
#> 70   1.187316390
#> 71  -1.783120154
#> 72  -0.187312263
#> 73  -0.587434154
#> 74  -0.501312836
#> 75  -0.880009002
#> 76   0.719143485
#> 77  -1.205695364
#> 78  -0.281889760
#> 79  -0.711889056
#> 80   0.845713134
#> 81   0.996642035
#> 82  -0.613218225
#> 83  -0.813743847
#> 84  -0.912547628
#> 85  -0.253357826
#> 86   0.521303642
#> 87   0.778124140
#> 88   0.013331095
#> 89   1.239880333
#> 90  -0.524634248
#> 91  -0.158703246
#> 92  -0.646916217
#> 93  -0.058358593
#> 94  -0.561692196
#> 95   1.360323813
#> 96   0.494979235
#> 97  -0.640262864
#> 98  -1.023159494
#> 99  -0.965794922
#> 100 -0.871369493