Convert a Summarized Experiment to a data.frame

This is a small helepr function to concatenate the colData and assays within SummarizedExperiment objects into a single data.frame. This is not necessary for the essential multimedia workflow, it is only exported for potential independent interest.

Usage

exper_df(exper)

Arguments

exper

An object of class SummarizedExperiment.

Value

A data.frame combining all slots of a multimedia object.

Examples

demo_joy() |>
    multimedia:::exper_df()
#>              ASV1        ASV2        ASV3        ASV4          ASV5 treatment
#> 1    1.4354692420  0.75246061 -0.19318124 -0.81954151  2.1687272723 Treatment
#> 2    0.1366364884 -0.01621241  0.88556168  0.81986097  0.5520854149 Treatment
#> 3   -0.0007021005  0.38926708 -1.56460893  0.37851509 -0.7804893154   Control
#> 4   -0.4885620001 -0.55089060  0.04741716 -0.60572474 -1.7947238906 Treatment
#> 5   -0.4194913681  0.63446208 -1.04270999 -0.75070896 -0.1711691617 Treatment
#> 6   -0.6635161639 -2.55640131  1.71864227 -1.17751891  0.7588329876 Treatment
#> 7    1.1379940488  0.70484405 -0.90610419  1.20664369  1.1636752552 Treatment
#> 8    2.2631410124  0.30460748  1.97810029  1.15096289  1.3353387195   Control
#> 9   -0.9416772527 -1.81836159 -1.03626091  0.75226329  0.1330150735 Treatment
#> 10  -0.2430331490 -0.33416466  0.92462868  0.35901625 -0.8621043421   Control
#> 11   0.8627872138  1.35870613  0.15298730 -0.96310416  0.1793946751 Treatment
#> 12   1.8091556156 -1.33360300 -1.27723010 -1.02945305 -1.1842059842 Treatment
#> 13  -0.3516685888 -0.65980196  0.68846451 -0.56212218 -0.2407523940 Treatment
#> 14  -1.5077575966 -1.36632874 -0.57906962  0.53235713 -0.7989827810 Treatment
#> 15   0.0810570875  0.64455756 -0.80791741  0.73938168 -1.1419811186   Control
#> 16   1.4721453069  0.12021896 -1.62597500 -1.57089704  0.9284590043 Treatment
#> 17   0.3278564189 -2.14044675  0.36118972 -0.83244312 -0.2547066147 Treatment
#> 18   0.5791297026  1.59164435 -0.31415824 -1.90421629 -0.1367485229 Treatment
#> 19  -1.0251035668  0.80146951 -0.63792473 -0.64205360  0.8960560591   Control
#> 20   0.4752507767  1.46858683 -0.48727922 -0.68058448  1.4246665706   Control
#> 21  -2.6375816009 -1.41718985  0.29596934 -0.35262779  0.5739132776 Treatment
#> 22   1.0978591914  0.47353608 -1.21652240 -0.53563705 -0.0367853234   Control
#> 23   1.1813595201 -0.89005096 -1.71720703  0.12669886  0.0085582686   Control
#> 24   1.7002915341 -0.35548540  2.15132630 -0.37395182  0.1621965517   Control
#> 25   0.0504436445  0.45177550  0.06908669 -0.03397225  1.8185346740   Control
#> 26  -0.2889850913  0.59782813 -0.90445636 -0.25851378 -1.4847277765 Treatment
#> 27  -0.6290483071 -0.08808419 -0.79161187  0.73647718 -0.8273039134   Control
#> 28  -0.6154138084  0.88706364 -0.89024037  0.88589250  0.8011981667   Control
#> 29   0.4509577935 -0.52662658  0.72848019 -0.34474319  0.8121667510 Treatment
#> 30   0.1782263483  0.43287846  0.63923671  1.99501840  0.1780518866   Control
#> 31  -0.7559110268  1.11130433 -0.90251632 -1.34534473 -0.3707174783   Control
#> 32   0.0330770114  0.48874380 -0.26701280  1.15850188 -0.9761119506   Control
#> 33   1.2670828333 -1.50621366  0.63961729 -0.02633289 -0.1407214126 Treatment
#> 34   1.8722198360  0.36628310 -0.74211673 -0.22629776  1.1382612183 Treatment
#> 35   0.0679279602  2.11998831  0.98200204  0.80083310 -1.6688172950   Control
#> 36   1.0256807141  1.23615048  0.55417986  0.11923133  1.6289451789   Control
#> 37   1.1847595803  1.48026835  0.05333653 -1.43489568 -2.2136142430   Control
#> 38   0.3184902178  0.35164811 -0.93794207  1.69841272 -0.5218286408   Control
#> 39   1.9662306736 -0.20596155  0.10708209 -1.10420470  1.2498740613 Treatment
#> 40   2.7914883776 -1.96984842  1.56611202  1.27552003  0.2277281438   Control
#> 41  -1.8504501961 -0.42444331  0.27360607  0.17236388  2.1330047748 Treatment
#> 42   0.5347560002  0.55972592 -0.04844703 -1.34991347  1.9332733615 Treatment
#> 43  -0.2577823316 -0.29963079 -0.12844856  2.48748452  1.2585113185   Control
#> 44   0.2793369721 -0.67193775 -0.19937084 -0.61593374 -1.9868656297   Control
#> 45   0.5271148380  0.41561420  0.66671604  0.39242172  0.2219983026 Treatment
#> 46  -0.0350131620 -0.51772139 -0.26673498 -0.99943973 -0.1725943525 Treatment
#> 47   0.0339399984 -0.46721671  1.85748344  0.38640701 -3.0881519922   Control
#> 48  -0.8310491834  1.33540829 -0.96055291  0.53018427  0.0005008653 Treatment
#> 49  -0.2084021276  2.05972135 -0.76806577  1.25154077  0.2591984935 Treatment
#> 50  -1.0026604770  0.36007963 -0.67118089  0.41071011  2.0127999944 Treatment
#> 51   0.3105846938  0.57324616  0.36933560 -1.08781195 -0.1623736283 Treatment
#> 52  -1.4437001210 -0.17266463 -0.20237016 -0.98441379  0.4766637042 Treatment
#> 53  -0.6600575034 -0.23177625  1.72024406  0.64862145  2.0252452707 Treatment
#> 54  -0.4191990796 -0.12208979  0.28213861 -0.88499669  1.3105182471 Treatment
#> 55  -0.2207604871  0.49615655 -0.14282793  0.33401513 -0.5283477983   Control
#> 56  -0.2491353424 -0.19734280 -0.71688387  0.21605796 -0.4131487349 Treatment
#> 57   0.0088186147  1.22456863 -1.17722933 -2.04281582 -0.4133461762 Treatment
#> 58  -0.6964503152 -0.79767447  0.81331498  2.39763101  0.2919063027 Treatment
#> 59  -0.4919798073  0.47296015  1.79571272  2.02862876  0.1078753882 Treatment
#> 60   0.3467045665 -0.26322148  1.31101366  0.90092549  0.9392112545   Control
#> 61   0.6441605332  0.31468826 -0.43706385 -0.45686729 -0.4356229027   Control
#> 62  -0.0060871536 -0.95927984 -0.14801404 -1.40635605 -1.0004008863   Control
#> 63   0.6139957525  2.78627300 -1.10825066  0.33363905  0.2994955267 Treatment
#> 64   0.2795901525  2.19252579  1.97362999  0.48909298 -0.3353482372   Control
#> 65  -1.1397064633 -1.59871696  2.09869967  0.75421355  0.1217997642 Treatment
#> 66   0.9260714028  0.20521114  0.44307997 -2.06963297  0.0530878761 Treatment
#> 67   2.1581168028  1.10005052  0.66320968  1.55310076  0.3086775207   Control
#> 68  -0.6637099512  2.18016354 -1.55458770  0.83030565 -1.1329613016 Treatment
#> 69   0.3654885787 -0.14321414  2.33448249 -1.66330787  0.7390708911 Treatment
#> 70  -0.4957029370 -0.09932915  0.44164445  0.47687835  0.0708298959 Treatment
#> 71  -0.0798237266  1.16408990  0.89782906 -0.23493671 -0.2368641120   Control
#> 72  -0.0347017637  0.03561307 -0.88526454 -0.44411404  1.9208670347   Control
#> 73  -1.8366257310  1.05664022 -0.61683445 -0.41228455  1.0652471318 Treatment
#> 74   0.5030445871 -0.85096977  1.11806658 -0.01198053  1.2336237287   Control
#> 75   0.9581452875 -0.26752721  1.20425244  0.72225629 -1.0325991430   Control
#> 76   0.6958681890  0.13690577  2.21570694 -0.28438372 -0.0877362984   Control
#> 77   0.2071444753  1.84164246 -1.58001358 -0.51915942  0.1074499799 Treatment
#> 78   1.1762604776  0.66605101 -0.62783591 -0.46993636  0.5152709971   Control
#> 79   1.6154410709  0.34560207 -0.53967391 -0.75695688 -1.5571952412   Control
#> 80  -0.4048665719 -0.10366260 -1.60524567 -0.55571660  0.8672501739 Treatment
#> 81  -0.8717413769 -0.70228188  1.02593576  0.71816434 -0.3988941921 Treatment
#> 82   0.4497730673  0.33879857  0.77100026 -0.54382271 -0.8429480374   Control
#> 83  -0.0935213882  0.13670219  0.37351629  0.53816650  0.1561003571   Control
#> 84  -0.6377318697  0.07491232 -0.15355540  0.32717711 -0.1716754051 Treatment
#> 85   0.0191600411  1.19496399 -1.26952820  1.33459201 -0.4659609899   Control
#> 86   0.1961793022  1.42529342  0.71265625  1.09730484 -0.8795958023   Control
#> 87   0.4720259679  0.72237982 -0.24710646  0.13087006 -0.5485113434 Treatment
#> 88   0.8075298839  0.44092616  0.55994859  0.91729091 -1.0048164997 Treatment
#> 89   1.0435391687 -0.98682920  0.59016549 -0.02430508 -0.1273804709 Treatment
#> 90   0.0341276795 -0.40177580  1.23608708 -0.19490015  0.4703092019   Control
#> 91  -0.5776738200  1.35426664  0.52659032 -1.24795118 -0.5497703015   Control
#> 92  -0.2669504184 -0.56450784 -1.00863818  0.07363590  0.9004055434 Treatment
#> 93  -0.6896227922  0.95287807  1.55958102  0.92335320  0.8545314203 Treatment
#> 94  -1.0142521056 -0.38839049  1.20442853  1.32532402 -1.1880837181   Control
#> 95  -0.3066462174 -0.79184009  0.39379603  1.88101749  0.5758979061   Control
#> 96   1.1323807234 -0.22481847 -0.40021213 -1.19067361 -1.0098475797   Control
#> 97   0.9954111437 -1.42139060 -1.97500057 -0.12321264  0.8460633123   Control
#> 98  -0.3279738185  1.99542562 -0.11683845 -0.54505049 -0.6666030797   Control
#> 99   1.1020969289  0.69932250 -0.51037911 -2.06500220  0.8557542689 Treatment
#> 100 -0.3780877896  1.06653580  0.64614490  0.05941876  0.6610819132   Control
#>              PHQ
#> 1    1.896493712
#> 2    0.868447518
#> 3   -1.025681688
#> 4    0.898102925
#> 5   -0.306672988
#> 6   -0.533606038
#> 7   -1.111781782
#> 8   -0.190867978
#> 9    0.050167357
#> 10  -0.403509856
#> 11   0.785493610
#> 12  -0.952619436
#> 13  -0.501073935
#> 14  -0.952255122
#> 15  -0.902389858
#> 16  -0.293200114
#> 17  -1.072157242
#> 18   0.366577034
#> 19  -0.659141372
#> 20  -1.326991499
#> 21   0.784043889
#> 22   0.488644560
#> 23   1.035494636
#> 24  -0.002753248
#> 25  -0.972982288
#> 26   0.723619871
#> 27  -1.618396097
#> 28   0.174681084
#> 29   0.231951599
#> 30   1.594942630
#> 31   1.001649101
#> 32  -0.291285746
#> 33  -1.836480184
#> 34  -0.426121589
#> 35  -1.437321397
#> 36  -0.263658245
#> 37  -1.438942063
#> 38   0.555655008
#> 39  -1.693412611
#> 40  -0.510382903
#> 41   0.104948754
#> 42   0.237517764
#> 43   0.043120422
#> 44   1.458972404
#> 45  -0.873313273
#> 46   0.010488831
#> 47   0.721154501
#> 48   0.601501588
#> 49  -1.035660843
#> 50   1.775742775
#> 51   0.071761077
#> 52   0.595039037
#> 53   0.356783326
#> 54   1.164052505
#> 55  -1.906488863
#> 56   1.008290035
#> 57  -1.209478864
#> 58  -1.637900089
#> 59  -0.626973031
#> 60  -0.484370439
#> 61   0.147264696
#> 62   2.596952758
#> 63  -0.759261572
#> 64   0.126698916
#> 65   0.653957723
#> 66  -0.287458782
#> 67  -1.808337016
#> 68   0.069689884
#> 69  -0.751523960
#> 70   0.614066389
#> 71   2.616118633
#> 72  -0.310197247
#> 73  -0.225556509
#> 74   0.895580204
#> 75   2.314944066
#> 76  -1.070841158
#> 77   0.123428339
#> 78   0.205772328
#> 79  -0.941252315
#> 80   0.845274417
#> 81  -1.333634125
#> 82  -2.050314328
#> 83  -0.689944634
#> 84  -0.408879321
#> 85   0.486114912
#> 86   0.415104613
#> 87  -0.439245206
#> 88  -0.997126147
#> 89  -0.718125379
#> 90   0.531715662
#> 91   2.713507671
#> 92   0.310425982
#> 93  -0.278492626
#> 94   0.889777247
#> 95  -0.059815110
#> 96   0.288509261
#> 97  -0.469084811
#> 98   0.302411585
#> 99  -2.730041826
#> 100 -0.614215108