Convert a Summarized Experiment to a data.frame

This is a small helepr function to concatenate the colData and assays within SummarizedExperiment objects into a single data.frame. This is not necessary for the essential multimedia workflow, it is only exported for potential independent interest.

Usage

exper_df(exper)

Arguments

exper

An object of class SummarizedExperiment.

Value

A data.frame combining all slots of a multimedia object.

Examples

demo_joy() |>
    multimedia:::exper_df()
#>             ASV1         ASV2        ASV3        ASV4         ASV5 treatment
#> 1    1.169159237  0.883856709  0.77061407  0.04634831  0.418207314 Treatment
#> 2    2.532692435 -0.844072722  0.04232179 -1.04645119  0.862889929   Control
#> 3    0.615294307 -1.945230019 -0.21052286  0.32640229 -0.308355670 Treatment
#> 4   -0.402409830 -1.706364591  2.55349296  0.01899164  0.248394505   Control
#> 5    1.171174751  0.769667594 -0.78147277 -1.09266635 -2.061460846 Treatment
#> 6    0.017835988  0.985268398  1.23226513 -0.29889042 -0.257470783   Control
#> 7   -0.989306571 -0.310184836  2.57303841  1.94052124 -0.334570167 Treatment
#> 8    1.374144148  0.463156537 -0.23874589 -0.63286310  0.096477705 Treatment
#> 9    0.001418294  0.353521990  0.63562556  1.08069937  0.589320162   Control
#> 10   0.055721827 -0.439919467 -0.32267614  0.46268597  0.532942303 Treatment
#> 11  -0.938377280  0.114921125 -0.04228148  0.46049300 -0.001039308 Treatment
#> 12   2.660793373 -0.372733979 -0.86378425  0.94708546 -0.083426968   Control
#> 13  -0.435271762 -1.755953474  1.95142894  1.87479662  0.180385437 Treatment
#> 14   2.187708587 -0.456596937  0.74256958  2.13397984  0.296437261   Control
#> 15   0.641642992  0.088507398 -0.40638154  2.10667876  0.944377087 Treatment
#> 16   0.837444927  0.518632617 -0.18907231  0.99573606  0.779806669   Control
#> 17  -0.481120267 -0.489330773 -1.33555095 -0.95078891 -0.780159242 Treatment
#> 18  -1.394575531 -1.256722074  1.12983595  0.48034140  1.349828551   Control
#> 19  -0.521954016 -1.456612346  1.49177211 -0.93160794  0.414151057   Control
#> 20  -0.955007383 -0.103840535  0.28657362  0.20152816  1.143908104   Control
#> 21   0.285110444  1.538836564 -0.33304578 -0.67620551 -1.393742222   Control
#> 22   1.027171421 -0.280624861  0.31854077  1.06022995 -0.251989311   Control
#> 23   1.636532488  0.836211008  0.77868510  0.95554666  1.984639433   Control
#> 24   1.545919703 -0.002319452 -0.26194718  0.33675879 -1.240723139 Treatment
#> 25  -0.238051997  1.656288367  1.57257058  1.20198633 -1.169361192 Treatment
#> 26   1.666936933  0.426075811  0.98546066  0.48994064  1.617361827   Control
#> 27  -0.958692481  0.055900287  0.33105430 -0.53852513 -0.169482822 Treatment
#> 28  -0.790513775 -0.628176399  0.62579871 -0.33558792  0.957219704   Control
#> 29  -1.549198481 -1.273103073  0.26449060 -1.39465593  0.257461984 Treatment
#> 30  -0.163217884 -1.177158800 -0.12173925 -0.46520259  0.388969148   Control
#> 31   0.970909051 -1.844861736  0.68074580  0.91881803 -0.225806072   Control
#> 32   0.155848497  0.092100311 -0.85625974 -0.35062084  0.476920717 Treatment
#> 33   0.762345838 -0.836123044 -0.86253459 -0.14685698 -0.107812638 Treatment
#> 34  -0.589230000 -0.440824299  2.39245005  0.08314347 -0.284215164 Treatment
#> 35   1.142656318 -0.948621004 -1.53075309  0.01872778 -0.573432234   Control
#> 36   1.661961996 -0.077077953  0.03557124  0.82506010 -0.575466663   Control
#> 37  -1.075269301  0.100251536  1.41887731 -0.06944164 -1.162914375 Treatment
#> 38   0.304582524 -0.708902603 -1.15464752  1.05561825  1.026694507 Treatment
#> 39  -0.658661765  1.338707008 -0.46449612  1.92734132 -0.060866173   Control
#> 40  -1.151540313  0.213164988  1.40556708 -3.06223138  0.637363776 Treatment
#> 41  -0.687828379 -0.143943267 -0.19471810 -0.56157555 -2.127263190 Treatment
#> 42  -1.458925882 -1.278664375 -0.03386924 -0.15539907  1.107334868   Control
#> 43  -0.600520305  1.021908941 -0.12826678 -1.30577694  0.824669143 Treatment
#> 44  -1.615579589 -0.155615356 -0.07461124  0.15903377  0.538736541   Control
#> 45  -0.813752419 -0.029268076  2.02430414  0.44844534 -0.663441408 Treatment
#> 46  -0.256523824 -0.671954228  0.92434064 -0.74098554 -2.348756365   Control
#> 47   1.280571384  0.690243775 -2.08335502 -1.20930462 -0.665739287   Control
#> 48   1.494572918  0.563802177 -0.02287051 -1.06032607  0.309429290   Control
#> 49  -1.141330289 -0.222086284  0.26747406 -0.30939629  0.372679088   Control
#> 50   1.215061047  1.486688387 -0.50450237 -0.67485940  0.419028642   Control
#> 51   0.120401035  2.499678765  0.88704756 -0.41835621 -0.743751640   Control
#> 52  -0.280815146  0.213880038 -0.76724189  1.54837747  1.413425575 Treatment
#> 53   0.193427026  0.660468194  0.04199718 -0.54787354  0.009850661   Control
#> 54   0.581687589 -1.280138476 -0.58862873 -0.90572204  0.259448884 Treatment
#> 55  -0.226053568  0.196512470 -1.42188807 -1.42412232 -1.585879684   Control
#> 56   0.457217205  0.236091147 -1.87588497 -0.37150698 -0.573477628 Treatment
#> 57   0.049710854 -2.416683984 -1.59592413  2.70184712 -0.752430828 Treatment
#> 58  -0.450067271  0.780391906 -1.29281762 -1.45222508 -0.049076189   Control
#> 59   0.076304827  0.520965255 -0.01595167  0.52084432 -0.361033475 Treatment
#> 60  -2.473944310 -0.350252238  0.71259061  1.09781138 -0.976658793 Treatment
#> 61  -0.150103070 -0.346022373  0.36355084 -0.97556000  0.946410076   Control
#> 62  -0.241641343 -0.718523362  1.59733887  1.35909953 -0.309943008 Treatment
#> 63   1.846481317  1.643756851 -0.46609799 -0.08590035  3.113403449   Control
#> 64   0.794904549  0.242316709 -0.51241441 -1.51450572 -0.963564214   Control
#> 65   0.563508970 -1.403109169  0.74963804 -0.96135069 -2.084951664   Control
#> 66   0.055949279 -0.665154824  0.19248435  0.95766351 -1.959474595 Treatment
#> 67   1.300089433 -0.816937674  2.11120792 -0.40194074  1.590038951   Control
#> 68   0.120833458  0.116326545  0.27072293 -0.53574330  0.159617515 Treatment
#> 69   0.148904500 -0.529874557  0.22623003  0.95372293 -2.275436756 Treatment
#> 70   0.519569526  0.171423564 -0.26647594  1.56097342 -1.533004675 Treatment
#> 71  -0.479681402  0.740178887 -0.14902531  0.12276728  0.875019796   Control
#> 72  -0.637082145 -0.667676862 -0.36649273  1.16862049  0.043243463   Control
#> 73   2.058002815 -0.427136823 -0.63275273 -0.83277801  1.305415246 Treatment
#> 74  -0.901310195  0.731458745  0.77969175 -0.07862134  0.281521137 Treatment
#> 75  -1.464061027  0.109826718 -0.59855327 -0.69787906 -0.900367662   Control
#> 76   0.158796857  0.728057895  0.24780634  0.86082750 -0.235792970   Control
#> 77  -0.347190745  0.348563397 -0.75255608  2.12287089 -0.654519576   Control
#> 78  -1.115136924  1.482109074 -1.28270211 -0.80413813  0.109878270 Treatment
#> 79   1.678149036 -0.458668825  0.56694515 -2.21378809  0.122738485 Treatment
#> 80   0.217740127 -0.809480614 -2.32099180  0.47716244  1.243062618 Treatment
#> 81   0.268284370  0.667790959  0.58795386  0.93687126  0.874638684   Control
#> 82  -0.894481988  1.087725543 -0.30202421  0.67472602 -0.116666619 Treatment
#> 83   1.260111821 -0.565971977 -1.35910630 -0.28310009  0.893152383   Control
#> 84   0.749487073 -0.236076588 -0.52937605  2.74850542  0.694941019 Treatment
#> 85   0.573994180  0.980337731  0.09033489  0.71780009 -0.703548968   Control
#> 86   0.794829743  0.027309279 -1.04615783  0.29668021  0.801820812 Treatment
#> 87   0.884796010 -0.182652856  0.06895633  0.27066702 -0.742361374 Treatment
#> 88   0.812692757  0.299793685 -0.92526207  0.01140376  1.008555927   Control
#> 89   0.740119935  0.941135467  1.78494670 -1.06017008  1.578248428 Treatment
#> 90  -0.139376293 -0.373464240 -0.24284473  0.97939272 -1.529947824   Control
#> 91   0.771712835 -0.431902289 -1.09531137  1.58136713  2.818283164 Treatment
#> 92  -0.148477118  0.586640784 -1.23467509  0.07276291 -1.357072112   Control
#> 93   0.946629697 -1.201894602  0.02106750 -0.87848925 -0.742942002   Control
#> 94  -0.412105029  1.381473771  1.12303239  0.58535798 -0.580790083 Treatment
#> 95  -0.773501869 -1.331932200 -0.28550993 -1.05126968  0.541502644 Treatment
#> 96  -0.578015585 -1.351756300  0.64040791 -1.26146756 -0.081337245   Control
#> 97  -0.699545716  1.408879698  1.55958110  1.41538266  2.338598983 Treatment
#> 98  -1.605267930 -0.626294481 -2.00328699  0.15212532  0.592095298   Control
#> 99   0.279273510 -0.099270949  0.12639653 -0.72398297  0.909480623   Control
#> 100 -0.138426167  3.159366034 -0.44232863  2.09072680 -0.741756191   Control
#>             PHQ
#> 1   -0.35012104
#> 2    1.31528376
#> 3    1.40800490
#> 4   -0.84204242
#> 5    0.78952472
#> 6    1.03699689
#> 7   -0.70983443
#> 8    0.09029496
#> 9    1.40042385
#> 10  -0.13600495
#> 11   2.16586523
#> 12   0.79958570
#> 13   1.08951115
#> 14  -0.77432117
#> 15  -1.31796067
#> 16  -0.09007568
#> 17   0.85875303
#> 18  -0.16530018
#> 19  -0.72523400
#> 20   1.26598756
#> 21  -0.06120078
#> 22  -0.31538145
#> 23  -0.66957208
#> 24  -1.50648836
#> 25   0.45189099
#> 26  -0.20282555
#> 27  -0.97292794
#> 28   1.35942543
#> 29   1.07966513
#> 30   0.34711918
#> 31  -0.21437071
#> 32  -0.17734785
#> 33  -0.42603376
#> 34   0.20090187
#> 35   1.67466816
#> 36  -0.09943760
#> 37   2.33032066
#> 38   0.95812246
#> 39  -1.06400887
#> 40   1.76065642
#> 41   0.66060539
#> 42   0.52736418
#> 43   1.17732341
#> 44   0.96148458
#> 45   0.82943848
#> 46  -1.37181788
#> 47  -0.21609445
#> 48  -2.28117150
#> 49  -0.33018782
#> 50  -0.43285741
#> 51   0.26626608
#> 52   0.75708659
#> 53  -0.42344689
#> 54   0.05188514
#> 55  -0.16768005
#> 56  -0.52934786
#> 57  -1.04865236
#> 58  -0.61597444
#> 59  -0.59254135
#> 60   1.35776645
#> 61   1.48113949
#> 62   0.21661891
#> 63   0.46200858
#> 64   0.37546713
#> 65   0.74372508
#> 66  -0.26852019
#> 67   0.46858847
#> 68  -0.42681778
#> 69   0.12161250
#> 70   2.01173972
#> 71   1.30660730
#> 72   0.15795449
#> 73   0.57289656
#> 74  -1.35250505
#> 75  -0.41358832
#> 76  -0.79451443
#> 77   0.08447398
#> 78   0.56089331
#> 79  -1.44519544
#> 80  -1.14505863
#> 81   0.05412823
#> 82   1.96791973
#> 83  -1.56473419
#> 84   0.81463816
#> 85   0.50636686
#> 86   0.06680147
#> 87   0.03376644
#> 88   0.39678133
#> 89  -0.86263238
#> 90  -1.37475980
#> 91   0.41048358
#> 92   0.52449749
#> 93  -0.73729998
#> 94  -0.99939775
#> 95  -0.29789631
#> 96   2.13253227
#> 97  -0.10328727
#> 98  -0.52814216
#> 99  -0.08012829
#> 100  0.23301333