Logistic Normal Multinomial Model

Apply a logistic normal multinomial model to jointly model a vector of count responses $y$ in an outcome or mediation model. This is a common choice for data where the parameter of interest is the composition across responses (e.g., microbiome).

lnm_model(...)

Arguments

...

Keyword parameters passed to lnm.

Value

model An object of class model with estimator, predictor, and sampler functions associated wtih a lienar model.

See also

Examples

m <- lnm_model()
mat <- data.frame(matrix(rpois(250, 10), 25, 10))
colnames(mat) <- paste0("y", 1:6)
fit <- m@estimator(y1 + y2 + y3 + y4 ~ y5 + y6, mat)
#> ------------------------------------------------------------ 
#> EXPERIMENTAL ALGORITHM: 
#>   This procedure has not been thoroughly tested and may be unstable 
#>   or buggy. The interface is subject to change. 
#> ------------------------------------------------------------ 
#> Gradient evaluation took 0.00017 seconds 
#> 1000 transitions using 10 leapfrog steps per transition would take 1.7 seconds. 
#> Adjust your expectations accordingly! 
#> Begin eta adaptation. 
#> Iteration:   1 / 250 [  0%]  (Adaptation) 
#> Iteration:  50 / 250 [ 20%]  (Adaptation) 
#> Iteration: 100 / 250 [ 40%]  (Adaptation) 
#> Iteration: 150 / 250 [ 60%]  (Adaptation) 
#> Iteration: 200 / 250 [ 80%]  (Adaptation) 
#> Success! Found best value [eta = 1] earlier than expected. 
#> Begin stochastic gradient ascent. 
#>   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes  
#>    100        -2007.408             1.000            1.000 
#>    200         -673.458             1.490            1.981 
#>    300         -329.745             1.341            1.042 
#>    400         -268.820             1.062            1.042 
#>    500         -274.971             0.854            1.000 
#>    600         -257.071             0.724            1.000 
#>    700         -277.428             0.631            0.227 
#>    800         -260.920             0.560            0.227 
#>    900         -253.948             0.501            0.073 
#>   1000         -249.741             0.452            0.073 
#>   1100         -246.520             0.354            0.070 
#>   1200         -254.087             0.158            0.063 
#>   1300         -246.048             0.058            0.033 
#>   1400         -241.020             0.037            0.030 
#>   1500         -257.278             0.041            0.033 
#>   1600         -250.753             0.037            0.030 
#>   1700         -259.327             0.033            0.030 
#>   1800         -243.091             0.033            0.030 
#>   1900         -260.457             0.037            0.033 
#>   2000         -247.767             0.040            0.033 
#>   2100         -245.203             0.040            0.033 
#>   2200         -258.387             0.042            0.051 
#>   2300         -239.869             0.047            0.051 
#>   2400         -238.789             0.045            0.051 
#>   2500         -263.186             0.048            0.051 
#>   2600         -240.404             0.055            0.067 
#>   2700         -242.107             0.052            0.067 
#>   2800         -238.820             0.047            0.051 
#>   2900         -252.505             0.046            0.051 
#>   3000         -271.522             0.048            0.054 
#>   3100         -241.741             0.059            0.070 
#>   3200         -239.646             0.055            0.070 
#>   3300         -240.458             0.047            0.054 
#>   3400         -248.518             0.050            0.054 
#>   3500         -242.511             0.043            0.032 
#>   3600         -242.089             0.034            0.025 
#>   3700         -241.888             0.033            0.025 
#>   3800         -248.490             0.035            0.027 
#>   3900         -238.543             0.033            0.027 
#>   4000         -240.566             0.027            0.025 
#>   4100         -241.950             0.015            0.009   MEDIAN ELBO CONVERGED 
#> Drawing a sample of size 1000 from the approximate posterior...  
#> COMPLETED. 
#> Finished in  0.3 seconds.