Logistic Normal Multinomial Model

Apply a logistic normal multinomial model to jointly model a vector of count responses $y$ in an outcome or mediation model. This is a common choice for data where the parameter of interest is the composition across responses (e.g., microbiome).

Usage

lnm_model(...)

Arguments

...

Keyword parameters passed to lnm.

Value

model An object of class model with estimator, predictor, and sampler functions associated wtih a lienar model.

See also

Examples

m <- lnm_model()
mat <- data.frame(matrix(rpois(250, 10), 25, 10))
colnames(mat) <- paste0("y", seq_len(6))
fit <- estimator(m)(y1 + y2 + y3 + y4 ~ y5 + y6, mat)
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 3.4e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.34 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100         -550.525             1.000            1.000
#> Chain 1:    200         -353.113             0.780            1.000
#> Chain 1:    300         -295.345             0.585            0.559
#> Chain 1:    400         -254.043             0.479            0.559
#> Chain 1:    500         -283.726             0.404            0.196
#> Chain 1:    600         -308.110             0.350            0.196
#> Chain 1:    700         -284.041             0.312            0.163
#> Chain 1:    800         -237.070             0.298            0.196
#> Chain 1:    900         -261.248             0.275            0.163
#> Chain 1:   1000         -256.841             0.249            0.163
#> Chain 1:   1100         -240.906             0.156            0.105
#> Chain 1:   1200         -233.507             0.103            0.093
#> Chain 1:   1300         -245.815             0.089            0.085
#> Chain 1:   1400         -237.689             0.076            0.079
#> Chain 1:   1500         -252.686             0.071            0.066
#> Chain 1:   1600         -249.185             0.065            0.059
#> Chain 1:   1700         -234.165             0.063            0.059
#> Chain 1:   1800         -246.376             0.048            0.050
#> Chain 1:   1900         -238.360             0.042            0.050
#> Chain 1:   2000         -236.145             0.041            0.050
#> Chain 1:   2100         -242.726             0.037            0.034
#> Chain 1:   2200         -243.954             0.035            0.034
#> Chain 1:   2300         -237.551             0.032            0.034
#> Chain 1:   2400         -231.966             0.031            0.027
#> Chain 1:   2500         -237.740             0.028            0.027
#> Chain 1:   2600         -243.381             0.029            0.027
#> Chain 1:   2700         -233.624             0.026            0.027
#> Chain 1:   2800         -241.941             0.025            0.027
#> Chain 1:   2900         -247.895             0.024            0.024
#> Chain 1:   3000         -243.607             0.025            0.024
#> Chain 1:   3100         -232.492             0.027            0.024
#> Chain 1:   3200         -235.791             0.028            0.024
#> Chain 1:   3300         -233.398             0.026            0.024
#> Chain 1:   3400         -236.033             0.025            0.024
#> Chain 1:   3500         -233.024             0.024            0.023
#> Chain 1:   3600         -237.777             0.023            0.020
#> Chain 1:   3700         -235.846             0.020            0.018
#> Chain 1:   3800         -251.771             0.023            0.018
#> Chain 1:   3900         -234.046             0.028            0.018
#> Chain 1:   4000         -238.012             0.028            0.017
#> Chain 1:   4100         -231.697             0.026            0.017
#> Chain 1:   4200         -235.373             0.026            0.017
#> Chain 1:   4300         -230.382             0.027            0.020
#> Chain 1:   4400         -233.546             0.027            0.020
#> Chain 1:   4500         -237.591             0.028            0.020
#> Chain 1:   4600         -230.583             0.029            0.022
#> Chain 1:   4700         -236.125             0.030            0.023
#> Chain 1:   4800         -235.401             0.024            0.022
#> Chain 1:   4900         -234.106             0.017            0.017
#> Chain 1:   5000         -234.203             0.016            0.017
#> Chain 1:   5100         -232.365             0.014            0.016
#> Chain 1:   5200         -234.763             0.013            0.014
#> Chain 1:   5300         -231.159             0.013            0.014
#> Chain 1:   5400         -235.785             0.013            0.016
#> Chain 1:   5500         -235.467             0.012            0.010
#> Chain 1:   5600         -238.931             0.010            0.010
#> Chain 1:   5700         -238.588             0.008            0.008   MEAN ELBO CONVERGED   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.
#> Warning: Pareto k diagnostic value is 1.86. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.