Logistic Normal Multinomial Model

Apply a logistic normal multinomial model to jointly model a vector of count responses $y$ in an outcome or mediation model. This is a common choice for data where the parameter of interest is the composition across responses (e.g., microbiome).

lnm_model(...)

Arguments

...

Keyword parameters passed to lnm.

Value

model An object of class model with estimator, predictor, and sampler functions associated wtih a lienar model.

See also

Examples

m <- lnm_model()
mat <- data.frame(matrix(rpois(250, 10), 25, 10))
colnames(mat) <- paste0("y", 1:6)
fit <- m@estimator(y1 + y2 + y3 + y4 ~ y5 + y6, mat)
#> ------------------------------------------------------------ 
#> EXPERIMENTAL ALGORITHM: 
#>   This procedure has not been thoroughly tested and may be unstable 
#>   or buggy. The interface is subject to change. 
#> ------------------------------------------------------------ 
#> Gradient evaluation took 9.4e-05 seconds 
#> 1000 transitions using 10 leapfrog steps per transition would take 0.94 seconds. 
#> Adjust your expectations accordingly! 
#> Begin eta adaptation. 
#> Iteration:   1 / 250 [  0%]  (Adaptation) 
#> Iteration:  50 / 250 [ 20%]  (Adaptation) 
#> Iteration: 100 / 250 [ 40%]  (Adaptation) 
#> Iteration: 150 / 250 [ 60%]  (Adaptation) 
#> Iteration: 200 / 250 [ 80%]  (Adaptation) 
#> Success! Found best value [eta = 1] earlier than expected. 
#> Begin stochastic gradient ascent. 
#>   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes  
#>    100         -514.928             1.000            1.000 
#>    200         -336.058             0.766            1.000 
#>    300         -256.892             0.613            0.532 
#>    400         -264.993             0.468            0.532 
#>    500         -245.380             0.390            0.308 
#>    600         -269.915             0.340            0.308 
#>    700         -269.146             0.292            0.091 
#>    800         -275.829             0.259            0.091 
#>    900         -239.006             0.247            0.091 
#>   1000         -236.131             0.224            0.091 
#>   1100         -251.013             0.129            0.080 
#>   1200         -239.288             0.081            0.059 
#>   1300         -238.181             0.051            0.049 
#>   1400         -265.414             0.058            0.059 
#>   1500         -237.154             0.062            0.059 
#>   1600         -260.400             0.062            0.059 
#>   1700         -279.481             0.068            0.068 
#>   1800         -267.812             0.070            0.068 
#>   1900         -237.202             0.068            0.068 
#>   2000         -236.610             0.067            0.068 
#>   2100         -238.169             0.061            0.068 
#>   2200         -239.703             0.057            0.068 
#>   2300         -248.231             0.060            0.068 
#>   2400         -234.876             0.056            0.057 
#>   2500         -238.727             0.045            0.044 
#>   2600         -243.598             0.038            0.034 
#>   2700         -247.348             0.033            0.020 
#>   2800         -236.277             0.033            0.020 
#>   2900         -239.703             0.022            0.016 
#>   3000         -247.244             0.025            0.020 
#>   3100         -234.995             0.029            0.030 
#>   3200         -237.821             0.030            0.030 
#>   3300         -237.970             0.026            0.020 
#>   3400         -243.538             0.023            0.020 
#>   3500         -234.867             0.025            0.023 
#>   3600         -236.298             0.024            0.023 
#>   3700         -242.453             0.025            0.025 
#>   3800         -237.619             0.022            0.023 
#>   3900         -235.641             0.022            0.023 
#>   4000         -250.560             0.024            0.023 
#>   4100         -234.105             0.026            0.023 
#>   4200         -233.190             0.025            0.023 
#>   4300         -232.394             0.026            0.023 
#>   4400         -235.391             0.025            0.020 
#>   4500         -237.693             0.022            0.013 
#>   4600         -236.553             0.022            0.013 
#>   4700         -235.192             0.020            0.010   MEDIAN ELBO CONVERGED 
#> Drawing a sample of size 1000 from the approximate posterior...  
#> COMPLETED. 
#> Finished in  0.4 seconds.