Logistic Normal Multinomial Model

Apply a logistic normal multinomial model to jointly model a vector of count responses $y$ in an outcome or mediation model. This is a common choice for data where the parameter of interest is the composition across responses (e.g., microbiome).

lnm_model(...)

Arguments

...

Keyword parameters passed to lnm.

Value

model An object of class model with estimator, predictor, and sampler functions associated wtih a lienar model.

See also

Examples

m <- lnm_model()
mat <- data.frame(matrix(rpois(250, 10), 25, 10))
colnames(mat) <- paste0("y", seq_len(6))
fit <- estimator(m)(y1 + y2 + y3 + y4 ~ y5 + y6, mat)
#> ------------------------------------------------------------ 
#> EXPERIMENTAL ALGORITHM: 
#>   This procedure has not been thoroughly tested and may be unstable 
#>   or buggy. The interface is subject to change. 
#> ------------------------------------------------------------ 
#> Gradient evaluation took 3.7e-05 seconds 
#> 1000 transitions using 10 leapfrog steps per transition would take 0.37 seconds. 
#> Adjust your expectations accordingly! 
#> Begin eta adaptation. 
#> Iteration:   1 / 250 [  0%]  (Adaptation) 
#> Iteration:  50 / 250 [ 20%]  (Adaptation) 
#> Iteration: 100 / 250 [ 40%]  (Adaptation) 
#> Iteration: 150 / 250 [ 60%]  (Adaptation) 
#> Iteration: 200 / 250 [ 80%]  (Adaptation) 
#> Success! Found best value [eta = 1] earlier than expected. 
#> Begin stochastic gradient ascent. 
#>   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes  
#>    100         -489.705             1.000            1.000 
#>    200         -335.284             0.730            1.000 
#>    300         -257.975             0.587            0.461 
#>    400         -288.378             0.466            0.461 
#>    500         -248.683             0.405            0.300 
#>    600         -245.982             0.339            0.300 
#>    700         -244.490             0.292            0.160 
#>    800         -243.673             0.256            0.160 
#>    900         -245.334             0.228            0.105 
#>   1000         -262.461             0.212            0.105 
#>   1100         -278.322             0.117            0.065 
#>   1200         -241.987             0.086            0.065 
#>   1300         -250.947             0.060            0.057 
#>   1400         -243.965             0.052            0.036 
#>   1500         -237.970             0.039            0.029 
#>   1600         -238.095             0.038            0.029 
#>   1700         -235.655             0.038            0.029 
#>   1800         -233.409             0.039            0.029 
#>   1900         -238.713             0.040            0.029 
#>   2000         -237.607             0.034            0.025 
#>   2100         -244.298             0.031            0.025 
#>   2200         -234.784             0.020            0.025 
#>   2300         -246.848             0.022            0.025 
#>   2400         -243.031             0.021            0.022 
#>   2500         -235.669             0.021            0.022 
#>   2600         -241.232             0.023            0.023 
#>   2700         -237.275             0.024            0.023 
#>   2800         -242.373             0.025            0.023 
#>   2900         -232.750             0.027            0.027 
#>   3000         -242.239             0.031            0.031 
#>   3100         -236.419             0.030            0.031 
#>   3200         -232.888             0.028            0.025 
#>   3300         -233.724             0.023            0.023 
#>   3400         -236.945             0.023            0.023 
#>   3500         -233.922             0.021            0.021 
#>   3600         -242.848             0.022            0.021 
#>   3700         -233.504             0.025            0.025 
#>   3800         -239.448             0.025            0.025 
#>   3900         -235.844             0.023            0.025 
#>   4000         -237.011             0.019            0.015 
#>   4100         -239.233             0.018            0.015 
#>   4200         -256.566             0.023            0.015 
#>   4300         -237.806             0.030            0.025 
#>   4400         -243.550             0.031            0.025 
#>   4500         -235.639             0.033            0.034 
#>   4600         -232.272             0.031            0.025 
#>   4700         -235.421             0.029            0.024 
#>   4800         -237.562             0.027            0.015 
#>   4900         -250.432             0.031            0.024 
#>   5000         -236.350             0.036            0.034 
#>   5100         -238.755             0.036            0.034 
#>   5200         -239.350             0.030            0.024 
#>   5300         -251.360             0.027            0.024 
#>   5400         -233.861             0.032            0.034 
#>   5500         -233.319             0.029            0.014 
#>   5600         -235.454             0.028            0.013 
#>   5700         -233.019             0.028            0.010 
#>   5800         -236.993             0.028            0.017 
#>   5900         -241.228             0.025            0.017 
#>   6000         -236.911             0.021            0.017 
#>   6100         -234.972             0.021            0.017 
#>   6200         -232.656             0.022            0.017 
#>   6300         -231.765             0.017            0.010 
#>   6400         -236.793             0.012            0.010 
#>   6500         -235.575             0.012            0.010 
#>   6600         -229.719             0.014            0.017 
#>   6700         -235.115             0.015            0.018 
#>   6800         -235.420             0.013            0.018 
#>   6900         -231.831             0.013            0.015 
#>   7000         -242.220             0.016            0.015 
#>   7100         -234.022             0.018            0.021 
#>   7200         -235.205             0.018            0.021 
#>   7300         -235.826             0.018            0.021 
#>   7400         -239.143             0.017            0.015 
#>   7500         -232.019             0.020            0.023 
#>   7600         -233.030             0.017            0.015 
#>   7700         -239.109             0.018            0.015 
#>   7800         -232.943             0.020            0.025 
#>   7900         -236.686             0.020            0.025 
#>   8000         -232.601             0.018            0.018 
#>   8100         -233.408             0.015            0.016 
#>   8200         -234.851             0.015            0.016 
#>   8300         -232.962             0.015            0.016 
#>   8400         -233.027             0.014            0.016 
#>   8500         -234.700             0.011            0.008   MEDIAN ELBO CONVERGED 
#> Drawing a sample of size 1000 from the approximate posterior...  
#> COMPLETED. 
#> Finished in  0.6 seconds.