Logistic Normal Multinomial Model

Apply a logistic normal multinomial model to jointly model a vector of count responses $y$ in an outcome or mediation model. This is a common choice for data where the parameter of interest is the composition across responses (e.g., microbiome).

Usage

lnm_model(...)

Arguments

...

Keyword parameters passed to lnm.

Value

model An object of class model with estimator, predictor, and sampler functions associated wtih a lienar model.

See also

Examples

m <- lnm_model()
mat <- data.frame(matrix(rpois(250, 10), 25, 10))
colnames(mat) <- paste0("y", seq_len(6))
fit <- estimator(m)(y1 + y2 + y3 + y4 ~ y5 + y6, mat)
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 9e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.9 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100         -481.839             1.000            1.000
#> Chain 1:    200         -301.516             0.799            1.000
#> Chain 1:    300         -288.004             0.548            0.598
#> Chain 1:    400         -263.187             0.435            0.598
#> Chain 1:    500         -279.370             0.359            0.094
#> Chain 1:    600         -256.310             0.315            0.094
#> Chain 1:    700         -261.325             0.272            0.090
#> Chain 1:    800         -268.773             0.242            0.090
#> Chain 1:    900         -248.363             0.224            0.082
#> Chain 1:   1000         -277.622             0.212            0.090
#> Chain 1:   1100         -247.618             0.124            0.090
#> Chain 1:   1200         -281.526             0.077            0.090
#> Chain 1:   1300         -261.070             0.080            0.090
#> Chain 1:   1400         -243.442             0.077            0.082
#> Chain 1:   1500         -247.382             0.073            0.082
#> Chain 1:   1600         -253.960             0.067            0.078
#> Chain 1:   1700         -253.519             0.065            0.078
#> Chain 1:   1800         -243.246             0.067            0.078
#> Chain 1:   1900         -260.596             0.065            0.072
#> Chain 1:   2000         -248.283             0.059            0.067
#> Chain 1:   2100         -248.740             0.048            0.050
#> Chain 1:   2200         -252.621             0.037            0.042
#> Chain 1:   2300         -243.268             0.033            0.038
#> Chain 1:   2400         -246.499             0.027            0.026
#> Chain 1:   2500         -252.644             0.028            0.026
#> Chain 1:   2600         -241.855             0.030            0.038
#> Chain 1:   2700         -247.926             0.032            0.038
#> Chain 1:   2800         -246.159             0.029            0.024
#> Chain 1:   2900         -241.776             0.024            0.024
#> Chain 1:   3000         -245.676             0.020            0.018
#> Chain 1:   3100         -246.302             0.020            0.018
#> Chain 1:   3200         -250.930             0.021            0.018
#> Chain 1:   3300         -244.292             0.020            0.018
#> Chain 1:   3400         -245.295             0.019            0.018
#> Chain 1:   3500         -243.305             0.017            0.018
#> Chain 1:   3600         -242.434             0.013            0.016
#> Chain 1:   3700         -239.725             0.012            0.011
#> Chain 1:   3800         -242.679             0.012            0.012
#> Chain 1:   3900         -243.447             0.011            0.011
#> Chain 1:   4000         -243.942             0.009            0.008   MEAN ELBO CONVERGED   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.
#> Warning: Pareto k diagnostic value is 1.16. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.