Access Class-Specific Mediators

This is a shorthand for accessing mediator-related slots in classes exported by the multimedia package.

Usage

mediators(object)

Arguments

object

An object whose mediators we want to access (either their names or values).

Value

The output will depend on the class of the object that is passed in. For multimedia objects, this returns a character vector of mediators. For mediation data objects, this returns the mediator mediator data.

Examples

exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
mediators(exper)
#>            ASV1        ASV2          ASV3        ASV4         ASV5
#> 1    0.53124145  0.11830678  0.0003225959  1.65134749 -0.415199557
#> 2    1.42293541  0.03704728 -0.6279917070  0.06685736  1.043990768
#> 3   -0.16803013  0.87058115  0.7906535111 -0.87122514  0.012441688
#> 4   -0.50520295 -1.16361621  1.3907377600  1.64378545  0.747379133
#> 5   -1.13180527 -0.16034861  0.8531901154  0.65046128 -2.290367543
#> 6   -1.02254694  1.07302430 -0.0927468480  0.29236298 -0.639362121
#> 7   -0.17535492 -0.31389494 -0.8300004089 -0.79485975 -0.592138986
#> 8   -0.45427752 -0.26747918 -1.1293656515  0.56028763 -1.678737124
#> 9   -1.39942670 -0.22227576 -0.7167226331  0.60343869 -0.170046068
#> 10  -0.75057719  0.23040670  0.9418321479 -1.11086524  1.287760725
#> 11   0.90854305  2.02735208 -1.2180628094  1.01259010  0.311343228
#> 12  -1.03504100 -0.25906236 -0.4884885344 -1.75196835  0.783292613
#> 13  -0.56405880  0.98445762  0.3131698036 -0.25967387  0.115242361
#> 14  -1.30947626 -0.61090274 -0.6200812627 -0.80497909 -0.962086155
#> 15   0.21578784  0.94564938  2.0352725283 -0.21372022 -1.758088976
#> 16  -0.12818148  0.37096115  1.5665822286  0.09517861 -0.799347610
#> 17  -0.02582579  0.98193428 -0.1248557494 -0.50712155  0.026926065
#> 18  -0.40866722 -0.71869261 -0.1199538276  0.92272902 -1.605597834
#> 19  -0.59622403 -2.46074715 -0.0687744288  1.05191635  2.210797217
#> 20   0.12856440  0.21275811  0.8020934919 -0.05798478  0.142671520
#> 21   1.08176434 -0.62923897 -0.9194905517  0.06477205 -0.791053130
#> 22  -1.38529266 -0.06363386 -0.4387457086 -1.51421377 -0.863314051
#> 23  -1.02506101  0.38835166 -1.1551399864 -1.50816017  0.595987972
#> 24   0.27280177 -0.90502918 -0.5353491861  0.11642028  0.301349163
#> 25   0.23387845 -0.03360864  0.1285469627  0.44905878  1.681518740
#> 26  -0.30921231  1.24746928  2.0783343233 -1.76256579  0.354261096
#> 27   0.40366656 -1.40134739  0.9964233784 -0.78791620  1.745230108
#> 28   0.28588958  0.44340178  0.2134951148 -0.46679177  0.214293738
#> 29  -0.22099915 -2.20945056 -1.0175478248 -1.06972814  1.168511515
#> 30  -0.19209530  1.22838478  0.8550478052 -0.23933339  0.914504024
#> 31   0.28868495  0.47096601  0.8687491266  1.05402556  1.742282736
#> 32  -0.39864466  0.22556722 -0.3851365154  0.87201342 -0.601995455
#> 33   0.55890935  1.28912919 -0.9258345554 -0.24144513  0.310670292
#> 34  -1.03404273 -0.53478454 -0.4169776954 -1.52017532 -0.185591452
#> 35   0.24788579  1.69767214  0.4145015441  1.11500797  1.727053843
#> 36   0.35782572 -0.62900092  0.5452459224 -0.52557097 -0.017184087
#> 37  -0.63754283 -1.26593183  2.7316177491  1.17447146 -1.334412199
#> 38  -0.16408133  0.09106318 -1.7820924181  0.51761366 -1.259858168
#> 39  -1.76393750  1.68144137  1.2561551603  1.79092693 -0.591677388
#> 40   0.21180952 -0.71435144  0.2952097470  0.60695908  0.724830520
#> 41  -1.76565138  0.74638789 -1.0759902363  0.51636654 -2.657927733
#> 42  -0.47068021  0.16827586  0.1315315796  0.75292871  1.757464953
#> 43  -0.53042396  0.10200676 -0.7827174646  1.29004184 -0.891255151
#> 44  -0.45157432 -0.83743054 -0.0488501810  0.80411238  0.737232665
#> 45   0.06536856  1.45499916 -1.3495887110  2.83024533  1.013020698
#> 46   1.99090501 -0.13161657  1.2327628985  1.96316314 -0.161006924
#> 47   0.39928125  2.07129980  0.2730937589  0.85816489 -0.193661538
#> 48  -0.24512583 -0.88872153 -0.0801801530 -0.37533173  1.160332931
#> 49  -1.31410697 -1.72536566  0.2301748601 -0.75780004 -0.848401105
#> 50  -0.17743564  1.64001767 -0.2727705356 -0.97155155 -1.441022245
#> 51  -1.31804228 -0.35432467  0.0233450592  0.38398272  0.009819443
#> 52   1.78121463  1.41544362  0.0713544851  0.80461405 -0.123563386
#> 53  -0.34416149  1.54014615  1.1735303347  0.34634615  0.702410920
#> 54   1.58426691  0.49049340 -0.5076742040  0.34982937  0.022047168
#> 55  -1.11568615  1.34239525 -0.3403594406  1.12469702 -0.370116055
#> 56   0.11558386 -0.99276374  0.0509881509 -0.06140264 -0.759887051
#> 57  -1.94082268  1.53894006 -0.6535780482 -1.23891764 -0.579172256
#> 58   0.20034076  0.51779112  1.8083958497 -0.04898333 -1.657712078
#> 59   0.99244430  0.76489392  0.8852078639  0.85744692 -0.644046194
#> 60   0.66519274  1.07189144  1.6766060205 -1.56450927  0.322941254
#> 61   0.16937271  0.62965667 -0.1975143054  0.93618718 -0.462517881
#> 62  -0.34139998 -2.06764967 -0.4198688261 -0.67161551  1.191134861
#> 63   1.18434488  2.04435545  0.7297626527 -0.21772416 -0.823833426
#> 64   0.86382885  0.08163495  1.2894244072 -0.89941398 -0.540336382
#> 65  -0.01801699 -0.99589681  0.7566018386  0.29371086 -1.460772373
#> 66   0.00440194 -0.44347772 -0.6801324074 -0.72504809 -1.105446773
#> 67   0.39157852 -0.58096627  0.7718645865 -1.35685903  0.240860245
#> 68   0.01883741  1.63023650  1.3204869074  0.17654361 -0.766428839
#> 69  -0.17752497 -3.25678565 -1.3429367841  0.77555063 -0.467725642
#> 70  -1.40764822  0.77188513 -1.5091572387  0.49813811  1.001875588
#> 71   1.23236035  0.24705334  1.4504478547  0.28527312  1.375720043
#> 72   0.43850083 -1.01235397 -2.0761891017  0.18327186 -0.260055802
#> 73  -0.62884721  0.28087716 -1.3942893928 -0.14086358  1.352580798
#> 74  -0.58781386  0.14975987  1.3573140603 -1.16176207  1.163082161
#> 75   0.89621702  0.29516967  0.4873417388 -0.37992727  1.653020047
#> 76   0.52606373 -0.42467631 -0.5920295347  0.98877715  1.081632362
#> 77   0.53971712  2.35658330 -1.2719071511  1.34528759  0.890678646
#> 78   0.74080031  0.12597369 -0.6451540350  1.78202435  0.415191555
#> 79  -0.68279913  0.39760422 -1.9901186360  0.08864263 -1.078266043
#> 80   1.28331594 -0.55727444 -0.8870747849 -1.53297286 -0.065153677
#> 81  -0.70375429  0.56056855 -0.2652842022 -0.57348270 -0.257673293
#> 82  -1.15064450  0.38697719 -1.2550639732 -1.41785164 -1.917494368
#> 83   0.41949846  0.21862042 -0.4222473752  0.69195173 -1.411220459
#> 84   0.10693213 -0.13988233 -0.3812729869  1.12464198  1.303718691
#> 85   0.17902858  0.69091008 -0.5943052967  0.87759794  0.358175633
#> 86   0.12968175  1.13798107  0.7112581525 -0.20267989 -0.050391812
#> 87  -1.80741237 -0.64984583 -0.1136991676  1.15587736 -0.580562926
#> 88   0.76792982  2.31932374  0.8207313046 -0.27386344  0.874646872
#> 89  -0.54254394 -0.51127174 -0.6660798039 -1.49238424 -0.458909476
#> 90  -0.48417598 -0.23503927  0.8209896660  0.23852295  0.138058815
#> 91  -0.36456058  0.44405696  0.2560220287  0.20800719 -0.793267069
#> 92  -2.12137151  0.14152728 -0.3536612294  0.31480984 -0.147815381
#> 93  -1.52671648 -0.67899724 -0.1127336824  1.58953211  0.634322722
#> 94  -0.36678625  0.82783326  1.0284644364  1.31220815 -1.853695550
#> 95   0.92413089 -1.08960547 -1.1464315455 -0.89487650 -0.265588470
#> 96   0.60951041  0.85743019  0.0816548401 -0.44562623  0.644190327
#> 97   0.19457802 -1.31194102  0.0314295676  1.31939366 -0.752161240
#> 98  -0.30542310  0.80549244 -0.0387929453  0.45133641 -2.883194421
#> 99  -1.02446182 -1.48186458 -1.0013369815  0.34222512  0.803116703
#> 100 -1.05362668 -0.51610137 -0.4919585904  2.38545950  0.439767316