Access Class-Specific Mediators

This is a shorthand for accessing mediator-related slots in classes exported by the multimedia package.

Usage

mediators(object)

Arguments

object

An object whose mediators we want to access (either their names or values).

Value

The output will depend on the class of the object that is passed in. For multimedia objects, this returns a character vector of mediators. For mediation data objects, this returns the mediator mediator data.

Examples

exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
mediators(exper)
#>            ASV1        ASV2        ASV3         ASV4         ASV5
#> 1    1.54019509 -0.02558980  0.39963708 -1.333433699  0.021485081
#> 2   -0.80435368 -0.04606819 -1.76997693  0.731350251  0.246063881
#> 3   -0.79785608  0.37187372 -0.22507957  0.312199194  1.411417326
#> 4    0.46419595  0.45097190 -0.88813508 -0.176704532  0.114792311
#> 5   -1.21043582  0.59785311  0.51499308  0.391564715 -0.873164510
#> 6    0.53481660 -3.27756666  0.43693473 -0.620287089 -1.065133834
#> 7   -0.93255534  0.48453409 -1.51466708  0.597327075 -0.841826837
#> 8   -0.91306020 -1.48068638 -0.40947950 -0.387601141  0.304128338
#> 9   -0.47055662 -0.95750106  0.46024422 -0.725498152 -0.849302397
#> 10   2.19389176 -0.02484174  0.62817828  0.560582890 -2.119540844
#> 11   0.30366884 -0.90970329  0.33550205  0.707172932  0.121781074
#> 12   0.75036961  2.33264018  0.25747284 -0.454687171 -0.938693500
#> 13  -0.43823145 -0.58038209 -1.15784611 -0.684070666 -0.274742034
#> 14   0.63185781  1.00929345  0.07206666  0.773891505 -1.471421412
#> 15  -1.98912342  1.75771524 -0.03727697 -0.095164486 -1.701907698
#> 16  -0.58215775 -0.06411622 -0.14170663 -1.432905361  1.307604406
#> 17  -0.57140764  1.20323173  0.84785272 -1.143526431  0.428479742
#> 18   0.53869546 -0.27685707 -2.60243200  0.387024693 -0.082944424
#> 19  -0.99024820  2.07755602 -0.17612760  0.660171946  0.473013413
#> 20  -1.58750974  0.37052253 -2.73916744  0.004910107  0.132156387
#> 21   0.22899502  0.33684901 -0.10622429 -1.102566958  0.357495159
#> 22   0.50304638 -0.49530427  0.65589292  0.772968270 -0.403215278
#> 23   1.23662699 -0.45752704  1.67730072  0.134935038 -0.405699453
#> 24  -0.35316237 -0.07506562 -1.08852195  0.392521121  0.578891737
#> 25  -1.17495938  0.30383052 -0.92480451 -0.700370241 -0.658339501
#> 26   0.39144160 -0.69777214 -0.18401555 -0.591544095  0.553331388
#> 27  -0.75709129  0.83630553 -1.51609318  0.918774036  0.015658518
#> 28  -0.22722292  0.69048642  2.16170128  0.594772916 -1.014764744
#> 29   1.75593887 -1.15549370  1.42067762  0.928408328  1.778325037
#> 30  -0.80087723  1.27640170  0.53194114  0.452616871 -0.481865746
#> 31   0.18531243 -1.79092721 -0.96083341  1.669757909 -1.505021843
#> 32   0.63176556 -0.07711113  1.01045636  1.479721466  0.315909929
#> 33   0.75567228  1.47477301  1.73314180 -1.187342529  0.741212602
#> 34   2.72017890 -0.12293075 -0.37501177 -0.842087596  0.288848393
#> 35   0.16381551  0.61127142  0.11057612  0.514602417 -0.324073709
#> 36   0.93847178  2.43219720 -0.54245997  0.975754432  0.200111491
#> 37   0.51198458  0.58817359 -0.56985525  0.160212710 -0.001279915
#> 38   3.19902020  0.79883033 -0.35887800 -0.441084656 -0.319263518
#> 39  -1.01657546  1.54692934 -1.62035693  2.394168242  0.772028614
#> 40   1.16740724  1.02229558 -0.25783876  0.293539852  0.489140024
#> 41  -0.25699881 -0.79592028  1.98413625 -1.848874522  0.489860525
#> 42  -1.28026011  0.39005757  1.01218156 -0.213154052  0.038402485
#> 43  -0.14494967 -1.20572239  0.10121674  0.093589295  1.211899070
#> 44  -0.13115271  0.08527883  0.45539204 -0.473575176 -0.673830666
#> 45   0.88168984 -0.45895716 -2.03157009  1.299604782  0.008896945
#> 46  -0.36346727 -1.49895530  1.15306517  0.520449428 -0.361230031
#> 47  -0.33956839 -0.25357590  0.35870339  1.747910962 -0.632387978
#> 48  -0.87606901 -0.08723383 -1.18242012 -0.495509770  0.578187494
#> 49  -1.40956242  0.06147626 -0.19474884 -0.492630529  0.497857487
#> 50   0.38056670  1.21790573 -0.70962877 -1.363284987  0.825420911
#> 51  -0.35702189  0.14846032  0.36834422 -0.832973900 -1.843216681
#> 52   0.52084749  0.50274779  0.98465475  0.589342590 -1.109841179
#> 53  -0.86772157 -0.21158441  2.31078278  1.846398062 -0.430856845
#> 54  -0.10221936  0.84390832  0.55082720 -1.349832292  0.313381822
#> 55  -0.27228771 -0.12012405  0.48219004  0.495750600 -0.686673269
#> 56   0.64658970  1.29909991  0.50331087 -2.004770181 -0.433256984
#> 57   0.97544418  1.33556703 -1.34213223  1.751768937  0.475813405
#> 58   0.15055712 -0.10991731 -0.08415801 -0.837222016  0.674862233
#> 59  -0.27884191 -1.34499016  0.64671317  1.061783446  0.231087839
#> 60   0.50899000 -0.21674215 -1.22907289  1.999380141  0.597313725
#> 61  -0.61645625 -1.12359685  1.83799782 -0.362827274 -0.194577932
#> 62  -2.13703329  0.52625029  0.78469057  1.880023668 -0.698958536
#> 63   0.41376821  0.08224441 -1.23200499 -1.182561174 -0.268634355
#> 64   0.22037235  0.28116092  0.15240585 -1.049491588  1.115374368
#> 65   0.71691026 -2.07680681 -0.54042393  1.059549049 -0.468408381
#> 66  -0.70649695  2.83328828 -0.69078682 -0.282516266  0.232551183
#> 67  -0.62120001  0.12774929  0.52526777 -0.287411504 -0.225307104
#> 68   1.77922485  0.27099994 -0.64712147 -0.504636131 -1.894228617
#> 69  -1.25385043 -0.78264350 -0.46621010  0.032247209 -1.847808836
#> 70  -1.74902096 -0.78307092  0.68554088 -1.292932905  0.269613187
#> 71   1.38435726 -0.81010756 -0.09422680  0.487110674  1.126303955
#> 72  -2.03722549  0.77849774 -1.71566021 -0.921443939  2.455605203
#> 73   0.25333670 -0.26087655  1.36773658  1.030503774  0.153879492
#> 74   0.57076815  2.01514292  0.05217894  1.259846415 -1.456029219
#> 75  -0.73875610 -0.50515666 -0.43548875 -1.783551156  0.168540649
#> 76   0.57583152 -0.27223356 -0.17635446  0.956051115 -0.679052778
#> 77   0.50087224  0.79493421 -0.06756084 -0.175083160  1.191097990
#> 78   2.35340628 -1.34120415 -0.72358357 -1.190187048 -0.283182750
#> 79   0.49857872 -0.83369624  0.24746791  1.216437117 -1.000208725
#> 80   0.93567221  0.65516512  1.01735988 -1.116391446  0.724202331
#> 81   0.09878434 -0.13082390 -1.65921608  1.074634879  1.187982942
#> 82   0.98608182 -1.43062407 -0.92725224 -1.265729276 -2.577124418
#> 83   0.34795665  0.26690084  1.67437745  0.367829058 -0.516852209
#> 84   0.13516469  0.94708070  0.31776888  1.596760204  0.428704927
#> 85  -0.79288021 -2.34255929  0.50509656  0.729149180 -0.153118544
#> 86   1.73670135  0.49975697  0.12461965  0.494194275  0.786190989
#> 87  -1.31794077  0.81075161 -0.69519951 -0.505556793  0.247334034
#> 88   1.72315132 -2.37419955 -2.19960487 -0.276520211  0.165915693
#> 89   1.17891271 -0.73002929 -0.06922245 -0.711807609 -0.134674002
#> 90   0.69394920 -0.42279129 -0.32772918  1.458877948  1.359610980
#> 91  -0.14992903  0.12442812  0.59559556  1.446418980 -1.303074447
#> 92   0.96198918  0.16720939 -1.09161495 -0.226713438 -0.826869217
#> 93   0.55582263  1.58418010 -0.67831461  0.540979230 -0.595212203
#> 94  -0.40060721 -0.09262031  0.92034602  1.074034525  0.006341155
#> 95  -1.06044409 -0.63711539  0.53688779  0.387444346 -0.397737226
#> 96   0.04231357  1.49745033  1.99288336 -0.870118217  1.473384704
#> 97  -1.30792213  1.26928428  0.04046978 -0.810849877  0.860567730
#> 98   2.05048309 -0.89121404  0.02699637 -0.680589162  0.947997276
#> 99  -1.37716383  0.20760376 -1.26706515  0.214132215  0.475210262
#> 100  0.03594102 -2.23986136 -0.09995620  0.381045227 -1.895832607