Access Outcomes

This is an getter method for outcomes in an S4 object, usually of class mediation_data.

Usage

outcomes(object)

Arguments

object

An object whose outcomes slot to modify.

Value

The output depends on the S4 class of the input object. If it is a multimedia model object, it will return a character vector of the outcome variable names. Otherwise it returns the outcome data.

Examples

exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
outcomes(exper)
#>             PHQ
#> 1   -0.08323820
#> 2   -0.89943319
#> 3    0.03699074
#> 4    0.38068084
#> 5   -0.19112915
#> 6   -0.04055868
#> 7    1.64619207
#> 8   -0.34516186
#> 9    1.01038338
#> 10  -0.34022125
#> 11   0.20334129
#> 12   0.79345976
#> 13  -0.21164766
#> 14   0.35688709
#> 15   0.02084397
#> 16  -2.05318345
#> 17   0.10112370
#> 18  -0.20965914
#> 19  -1.40774327
#> 20   0.74973582
#> 21   0.79731695
#> 22  -0.63565659
#> 23   0.47643922
#> 24   1.88500404
#> 25   0.74732153
#> 26   0.45776935
#> 27   0.09305775
#> 28   1.12694588
#> 29   2.18224585
#> 30   1.66312383
#> 31   1.74791358
#> 32   1.14287462
#> 33   0.16839946
#> 34   1.59694852
#> 35   0.68405984
#> 36   0.21656825
#> 37   0.24391366
#> 38   0.78259268
#> 39  -0.89670577
#> 40  -1.02955217
#> 41  -0.67009576
#> 42   1.49217025
#> 43   1.33146975
#> 44  -0.30029241
#> 45  -1.66098654
#> 46   1.32143666
#> 47  -1.90373537
#> 48  -0.64120132
#> 49   2.08218121
#> 50   1.30080053
#> 51   0.37433515
#> 52   0.35987360
#> 53  -2.33921779
#> 54  -1.55404951
#> 55   1.60976419
#> 56   0.54145044
#> 57   0.21535824
#> 58  -0.17543866
#> 59  -0.88690198
#> 60  -1.19801010
#> 61  -0.06281222
#> 62   0.06600860
#> 63  -0.58550860
#> 64  -0.01339130
#> 65   0.46095548
#> 66   0.38153469
#> 67  -0.99000229
#> 68  -0.13758227
#> 69   1.19014647
#> 70   0.22270007
#> 71  -0.16223296
#> 72   0.15387357
#> 73   1.63474120
#> 74   1.03233154
#> 75   0.53111644
#> 76   0.04127713
#> 77   0.37903994
#> 78  -2.56499730
#> 79  -0.03404557
#> 80  -0.29393063
#> 81  -2.56763714
#> 82  -1.88213077
#> 83   1.49026159
#> 84   1.22333119
#> 85   0.87819870
#> 86   1.45347379
#> 87   1.43067258
#> 88   1.59663619
#> 89  -0.55728350
#> 90   2.42572551
#> 91   0.01569235
#> 92   0.76348183
#> 93   0.31073558
#> 94  -0.93700269
#> 95   2.07183304
#> 96  -0.84888767
#> 97   0.37553626
#> 98  -0.15740186
#> 99   0.78034490
#> 100 -1.08525313