Access Outcomes

This is an getter method for outcomes in an S4 object, usually of class mediation_data.

Usage

outcomes(object)

Arguments

object

An object whose outcomes slot to modify.

Value

The output depends on the S4 class of the input object. If it is a multimedia model object, it will return a character vector of the outcome variable names. Otherwise it returns the outcome data.

Examples

exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
outcomes(exper)
#>              PHQ
#> 1    1.459119629
#> 2    1.075017006
#> 3   -0.648621400
#> 4   -0.766817724
#> 5    0.333862604
#> 6   -0.831113272
#> 7   -0.030076822
#> 8    0.253493696
#> 9   -1.011658235
#> 10   1.118257161
#> 11   1.765568777
#> 12   2.125882139
#> 13  -1.128714403
#> 14  -1.346916082
#> 15  -0.043135937
#> 16  -1.237819727
#> 17   1.040033592
#> 18  -1.753599344
#> 19  -2.156437886
#> 20   0.115867355
#> 21  -1.192660958
#> 22   0.545771147
#> 23  -1.115846087
#> 24   0.645905983
#> 25  -1.098728377
#> 26  -0.529053795
#> 27   0.021829580
#> 28   0.509103506
#> 29   0.797596727
#> 30  -0.418227153
#> 31  -0.174636326
#> 32   1.935767246
#> 33   0.054904936
#> 34  -1.703423039
#> 35   0.252985039
#> 36  -0.385277339
#> 37  -0.836604637
#> 38   1.348129876
#> 39  -0.257863504
#> 40  -0.845709977
#> 41  -1.163330148
#> 42   2.298011654
#> 43   1.139699445
#> 44  -1.056648130
#> 45   2.209536982
#> 46  -0.317008571
#> 47  -0.288019858
#> 48   0.611293181
#> 49   1.057959077
#> 50   1.523221912
#> 51  -0.146052841
#> 52   1.388064428
#> 53   0.442766299
#> 54   1.341311855
#> 55   0.573083286
#> 56   0.528052317
#> 57  -0.301344681
#> 58  -0.109083221
#> 59  -0.731855853
#> 60  -0.005858757
#> 61   1.717172454
#> 62   1.432206268
#> 63  -0.899772356
#> 64   1.045503952
#> 65   0.951350210
#> 66  -0.651139357
#> 67  -0.268392785
#> 68  -0.701613110
#> 69   0.739066684
#> 70  -0.710090876
#> 71  -0.719461346
#> 72   0.486113340
#> 73   1.386092016
#> 74  -0.114695376
#> 75  -0.599351537
#> 76  -0.806949952
#> 77  -0.210062255
#> 78  -0.071945326
#> 79   1.106198913
#> 80   0.467307940
#> 81   0.050910458
#> 82  -0.144065802
#> 83   1.444311937
#> 84  -0.339979483
#> 85   0.762980925
#> 86   1.199534754
#> 87   0.038637503
#> 88  -0.597283389
#> 89  -1.443494598
#> 90   0.806359565
#> 91  -0.011603784
#> 92   1.062121267
#> 93  -1.909071751
#> 94  -0.598972151
#> 95  -0.735415278
#> 96  -0.302541195
#> 97  -0.563486599
#> 98   1.583454135
#> 99  -0.380822843
#> 100  0.298767071