Outcomes Data in a Mediation Data Object

This is an accessor function to the @outcomes slot in a mediation data object. It returns the entire outcomes dataset, in contrast to outcomes() applied to a multimedia object, which only returns the names of the outcome variables.

# S4 method for class 'mediation_data'
outcomes(object)

Arguments

object

An object of class mediation_data.

Value

A data.frame whose rows are samples and columns different outcomes.

Examples

exper <- demo_joy() |>
  mediation_data("PHQ", "treatment", starts_with("ASV"))
outcomes(exper)
#>              PHQ
#> 1    0.550102149
#> 2   -0.602778290
#> 3   -0.559073604
#> 4   -0.915208331
#> 5   -0.140940208
#> 6    0.622742753
#> 7    1.710095545
#> 8   -0.210557670
#> 9   -0.833156569
#> 10  -0.399705381
#> 11  -1.110015495
#> 12  -2.346280355
#> 13   0.891904111
#> 14   1.542447032
#> 15  -0.280646712
#> 16   0.013551550
#> 17   1.181908893
#> 18   0.574661733
#> 19   0.169726522
#> 20  -1.105094421
#> 21  -0.835822421
#> 22   1.736278769
#> 23  -1.579294510
#> 24   0.464005514
#> 25  -0.895538088
#> 26   0.093038942
#> 27  -1.238394346
#> 28  -0.254344298
#> 29  -0.310976146
#> 30   1.616318967
#> 31  -1.093436240
#> 32   1.796682515
#> 33   1.329490539
#> 34   0.647924470
#> 35   0.862620339
#> 36  -0.194212743
#> 37   0.163319427
#> 38   0.808930248
#> 39   0.431651261
#> 40  -0.008146953
#> 41   0.042558717
#> 42   0.304229662
#> 43   1.190745453
#> 44   0.955126994
#> 45   0.960418001
#> 46   0.167526910
#> 47   0.354014144
#> 48  -1.979939148
#> 49   0.709426496
#> 50   1.775245234
#> 51   1.544155122
#> 52  -1.249423739
#> 53   0.214931730
#> 54  -0.617309310
#> 55   0.840174720
#> 56   0.110901898
#> 57   0.313872015
#> 58  -0.905026609
#> 59  -1.246358017
#> 60   1.696952331
#> 61  -0.781661994
#> 62   0.364998352
#> 63   0.849716594
#> 64   0.589325235
#> 65   0.534445297
#> 66   0.606456410
#> 67   0.489643003
#> 68   0.003074045
#> 69   1.662479391
#> 70  -0.038613351
#> 71  -0.183110148
#> 72  -0.163563214
#> 73  -0.660431071
#> 74  -0.763967709
#> 75   0.324635661
#> 76  -0.445772031
#> 77  -1.198121274
#> 78  -1.452291391
#> 79   0.947255638
#> 80  -0.321261508
#> 81  -0.740805739
#> 82   0.728885459
#> 83   0.204191518
#> 84  -0.691981375
#> 85   1.521127679
#> 86   0.238875612
#> 87  -0.014173865
#> 88  -0.428663510
#> 89   0.408272326
#> 90  -0.066179537
#> 91   2.112490689
#> 92  -2.483533303
#> 93  -1.174604744
#> 94  -0.683334177
#> 95   0.702581419
#> 96   0.081287765
#> 97   1.307252735
#> 98  -0.515860825
#> 99  -1.159720784
#> 100  0.399729689