Outcomes Data in a Mediation Data Object

This is an accessor function to the @outcomes slot in a mediation data object. It returns the entire outcomes dataset, in contrast to outcomes() applied to a multimedia object, which only returns the names of the outcome variables.

Usage

# S4 method for class 'mediation_data'
outcomes(object)

Arguments

object

An object of class mediation_data.

Value

A data.frame whose rows are samples and columns different outcomes.

Examples

exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
outcomes(exper)
#>              PHQ
#> 1   -1.050672134
#> 2    0.135282734
#> 3    0.051633446
#> 4    1.024384005
#> 5    0.226539885
#> 6    0.518594329
#> 7   -1.610305319
#> 8   -1.965779365
#> 9   -0.127218529
#> 10  -1.338718382
#> 11  -0.687794839
#> 12  -0.661524866
#> 13   0.201347721
#> 14   0.088391657
#> 15  -0.250645229
#> 16  -1.577244279
#> 17   1.890477573
#> 18  -0.001427814
#> 19  -1.695060417
#> 20  -0.322939427
#> 21   0.323506172
#> 22   0.584431943
#> 23  -1.206699412
#> 24   0.508678034
#> 25  -1.108005536
#> 26   0.945770012
#> 27   0.552432347
#> 28   0.209806501
#> 29   0.868112620
#> 30   0.480677868
#> 31  -0.935879727
#> 32  -0.122510436
#> 33  -0.781037858
#> 34   0.978406885
#> 35   1.053999570
#> 36  -2.258446462
#> 37   1.333816299
#> 38   1.209017000
#> 39   0.284511946
#> 40   2.164272229
#> 41   1.640966354
#> 42   0.492742078
#> 43   1.396644602
#> 44   0.991564706
#> 45  -1.092226622
#> 46  -0.983258458
#> 47   0.151518921
#> 48  -1.947338639
#> 49   1.830720952
#> 50   0.385542467
#> 51   0.608927499
#> 52  -0.134095218
#> 53   1.364287689
#> 54   1.573913615
#> 55  -0.392772530
#> 56   0.826509319
#> 57  -0.138094876
#> 58  -1.642060590
#> 59  -0.665453880
#> 60  -0.147753888
#> 61  -0.915016538
#> 62  -0.404575589
#> 63  -0.457153416
#> 64  -1.140807417
#> 65  -0.668572338
#> 66   0.692174161
#> 67  -0.844409334
#> 68  -0.572509566
#> 69   2.271149318
#> 70  -0.303002289
#> 71  -0.272201723
#> 72  -1.031123210
#> 73   0.412910169
#> 74   1.362841807
#> 75  -0.385824515
#> 76   0.712344919
#> 77  -0.091297194
#> 78   0.184083751
#> 79   1.113072744
#> 80  -0.386195937
#> 81  -0.285794262
#> 82  -0.347298938
#> 83  -0.224824113
#> 84   1.049400105
#> 85  -1.289021764
#> 86  -1.771555636
#> 87   1.844881210
#> 88  -1.395816760
#> 89   0.979955034
#> 90  -1.872677286
#> 91  -0.493231503
#> 92   2.726109910
#> 93  -0.268415996
#> 94   0.409785270
#> 95   0.863320966
#> 96  -1.051224013
#> 97   0.210666246
#> 98   1.599839288
#> 99  -0.415730406
#> 100  0.279960793