Outcomes Data in a Mediation Data Object

This is an accessor function to the @outcomes slot in a mediation data object. It returns the entire outcomes dataset, in contrast to outcomes() applied to a multimedia object, which only returns the names of the outcome variables.

Usage

# S4 method for class 'mediation_data'
outcomes(object)

Arguments

object

An object of class mediation_data.

Value

A data.frame whose rows are samples and columns different outcomes.

Examples

exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
outcomes(exper)
#>              PHQ
#> 1   -0.267923808
#> 2   -1.871058195
#> 3   -0.132344794
#> 4   -1.064807923
#> 5    0.381561497
#> 6   -0.399597986
#> 7   -0.687715176
#> 8    1.398099725
#> 9   -0.801692754
#> 10   0.797991203
#> 11  -0.803233624
#> 12   0.970691479
#> 13   0.002084574
#> 14   0.116555269
#> 15  -1.137732766
#> 16   1.103933367
#> 17  -0.222138858
#> 18  -0.833048051
#> 19   1.105513477
#> 20  -1.408871022
#> 21  -0.614001324
#> 22  -1.011274092
#> 23   0.487434169
#> 24  -0.008300809
#> 25   0.784329206
#> 26   0.430313069
#> 27   1.293212664
#> 28  -0.149974073
#> 29   0.146001843
#> 30  -1.356398087
#> 31   0.142518756
#> 32  -0.141323475
#> 33  -0.266015111
#> 34  -0.007484481
#> 35   0.663004857
#> 36  -0.862119664
#> 37   0.113170874
#> 38   2.240951587
#> 39  -0.498007068
#> 40  -1.193525697
#> 41   0.361465175
#> 42  -0.180280191
#> 43   0.456467347
#> 44  -1.937440080
#> 45  -2.855109598
#> 46   0.520609333
#> 47   1.723475191
#> 48   0.961833751
#> 49  -0.597092941
#> 50  -1.237158164
#> 51   0.094383976
#> 52   0.314298499
#> 53  -0.992155359
#> 54   0.855591845
#> 55   0.981075346
#> 56  -0.885570853
#> 57  -0.520414724
#> 58   0.334489212
#> 59  -0.432578690
#> 60  -1.207544322
#> 61   0.127247171
#> 62   0.886883253
#> 63   0.173009517
#> 64  -0.069385388
#> 65  -1.024063967
#> 66   0.461554491
#> 67   0.511996898
#> 68   0.597377565
#> 69  -0.199138490
#> 70   0.503500811
#> 71  -1.383427666
#> 72   1.023787416
#> 73  -0.201490896
#> 74  -0.502492131
#> 75  -1.719199699
#> 76   0.899701186
#> 77   1.765319680
#> 78   0.662786122
#> 79   1.083202734
#> 80   1.464910999
#> 81  -0.297582465
#> 82   0.403374918
#> 83   1.192714585
#> 84  -0.970655949
#> 85   0.053221507
#> 86  -0.041814201
#> 87   0.829981005
#> 88   0.005835926
#> 89   0.217809097
#> 90   2.641549115
#> 91   0.862678529
#> 92   0.379454682
#> 93  -0.762062947
#> 94  -1.388500220
#> 95  -1.312762301
#> 96  -0.267531936
#> 97  -2.309733000
#> 98   0.902255030
#> 99   1.041792680
#> 100  1.618068410