Outcomes Data in a Mediation Data Object

This is an accessor function to the @outcomes slot in a mediation data object. It returns the entire outcomes dataset, in contrast to outcomes() applied to a multimedia object, which only returns the names of the outcome variables.

Usage

# S4 method for class 'mediation_data'
outcomes(object)

Arguments

object

An object of class mediation_data.

Value

A data.frame whose rows are samples and columns different outcomes.

Examples

exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
outcomes(exper)
#>             PHQ
#> 1   -1.10238443
#> 2    0.08267221
#> 3    0.31937981
#> 4    0.32736982
#> 5   -0.58758760
#> 6    1.00126373
#> 7    0.16208593
#> 8    0.46262540
#> 9    1.63887589
#> 10   0.79999094
#> 11  -0.57216907
#> 12  -1.51631223
#> 13   1.14221807
#> 14  -1.19928498
#> 15  -1.12681603
#> 16   0.16378743
#> 17  -0.56466321
#> 18  -0.91070151
#> 19  -0.19147784
#> 20   1.76758423
#> 21  -1.23660768
#> 22  -0.55666110
#> 23  -1.44582560
#> 24   1.95324252
#> 25   0.21183626
#> 26  -0.27821589
#> 27   2.38156351
#> 28  -0.34841247
#> 29   1.32885014
#> 30   0.30819211
#> 31  -1.41062994
#> 32  -0.59043266
#> 33   0.40905734
#> 34  -0.50337821
#> 35   0.22147282
#> 36   0.13342470
#> 37  -0.54151276
#> 38   0.44310204
#> 39   1.70804289
#> 40   1.17596421
#> 41  -0.84081548
#> 42   0.18030677
#> 43   0.64803408
#> 44  -1.74315280
#> 45  -2.48059977
#> 46  -0.64985153
#> 47  -0.25999777
#> 48   0.10729246
#> 49   0.64600282
#> 50  -0.74144290
#> 51  -0.24269928
#> 52  -1.23928568
#> 53   1.64818673
#> 54   1.37771671
#> 55  -1.65980980
#> 56   0.15874353
#> 57  -1.85947685
#> 58   0.55305879
#> 59  -0.70444146
#> 60   0.14670569
#> 61  -0.52304329
#> 62  -1.00730298
#> 63   0.15937458
#> 64  -1.01000824
#> 65  -0.13920439
#> 66   2.16942307
#> 67  -2.02641745
#> 68  -0.81398357
#> 69   1.04692183
#> 70   0.68125514
#> 71   1.77640624
#> 72   1.17024674
#> 73   1.57346343
#> 74   1.10179397
#> 75  -1.08801944
#> 76  -0.76015673
#> 77  -1.78063256
#> 78   0.83237982
#> 79   1.01608872
#> 80  -0.30132148
#> 81   0.56339893
#> 82   0.97121031
#> 83   0.45078387
#> 84  -0.45854492
#> 85   0.71826860
#> 86   1.48603201
#> 87  -0.67492058
#> 88   1.12765095
#> 89   0.51523004
#> 90  -0.94574215
#> 91  -0.62690086
#> 92   0.69496343
#> 93   0.94589009
#> 94  -0.38837591
#> 95  -0.24684507
#> 96  -1.74522605
#> 97  -0.74811822
#> 98   0.67162492
#> 99   1.08675279
#> 100  1.12196847