Sample New Mediator/Outcome Data

This generalizes the built-in sample method to the multimedia class. Given an estimated multimedia object, this function supports sampling along the estimated DAG. It first samples M* | T, X and then Y* | M*, T, X. Each sampling step will call the sample method within the mediation and outcome models that make up the multimedia object on which this is called.

# S4 method for class 'multimedia'
sample(x, size, pretreatment = NULL, profile = NULL, mediators = NULL, ...)

Arguments

x

An object of class multimedia containing the estimated mediation and outcome models whose mediation and outcome samples we want to obtain.

size

A placeholder argument to agree with the default sample method in R base. We always return the number of samples as set in either the original input x or a new input profile.

pretreatment

By default, we will return mediation and outcome model predictions using the same pretreatment variables as used when initially estimating the models (like setting newdata = NULL in usual predict). To pass in different pretreatment variables, provide a data.frame here whose columns match the pretreatments as the originally trained mediation and outcome models.

profile

An object of class treatment_profile containing the treatment profile to consider in the difference. Defaults to a profile with all the unique treatment configurations observed in the original data, shared across both the mediators and outcomes.

mediators

By default, we will return outcome predictions using the predicted mediators from the mediation model. Modify this argument if you would like to directly control the mediation inputs for the outcome model. Must be a data.frame whose columns are named to match the mediators(object).

...

Additional options to pass to the @sampler method in the estimated mediation model.

Examples

exper <- demo_spline(tau = c(2, 1)) |>
  mediation_data(starts_with("outcome"), "treatment", "mediator")
fit <- multimedia(exper) |>
  estimate(exper)
samples <- sample(fit)
mediators(samples)
#> # A tibble: 2 × 1
#>   mediator
#>      <dbl>
#> 1    -5.66
#> 2     2.50
outcomes(samples)
#> # A tibble: 2 × 2
#>   outcome_1 outcome_2
#>       <dbl>     <dbl>
#> 1      2.74    -1.82 
#> 2      1.35     0.763

# sampling with just different "n" has no effect.
samples <- sample(fit, 100)
#> Warning: The size argument in sample is ignored. Please adjust the treatment profile to
#> adjust the number of samples.

# Instead sample at a new treatment configuration
t1 <- data.frame(treatment = factor(rep(c(0, 1), each = 50)))
profile <- setup_profile(fit, t_mediator = t1, t_outcome = t1)
samples <- sample(fit, profile = profile)
mediators(samples)
#> # A tibble: 100 × 1
#>    mediator
#>       <dbl>
#>  1   -1.76 
#>  2   -0.186
#>  3   -1.98 
#>  4   -4.35 
#>  5   -4.43 
#>  6   -3.58 
#>  7   -0.635
#>  8   -2.70 
#>  9    0.789
#> 10   -1.48 
#> # ℹ 90 more rows
outcomes(samples)
#> # A tibble: 100 × 2
#>    outcome_1 outcome_2
#>        <dbl>     <dbl>
#>  1   -0.0550    0.132 
#>  2    0.177     0.950 
#>  3    0.960    -0.0649
#>  4    1.25     -0.560 
#>  5    1.79     -0.181 
#>  6    0.531    -0.859 
#>  7    0.595    -0.205 
#>  8    0.402    -0.0511
#>  9   -0.104    -0.0459
#> 10    0.257    -0.406 
#> # ℹ 90 more rows